Network.AWS.DirectConnect

Description

AWS Direct Connect links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. With this connection in place, you can create virtual interfaces directly to the AWS cloud and Amazon Virtual Private Cloud, bypassing Internet service providers in your network path.

Documentation

module Network.AWS.DirectConnect.AllocateConnectionOnInterconnect

module Network.AWS.DirectConnect.AllocatePrivateVirtualInterface

module Network.AWS.DirectConnect.AllocatePublicVirtualInterface

module Network.AWS.DirectConnect.ConfirmConnection

module Network.AWS.DirectConnect.ConfirmPrivateVirtualInterface

module Network.AWS.DirectConnect.ConfirmPublicVirtualInterface

module Network.AWS.DirectConnect.CreateConnection

module Network.AWS.DirectConnect.CreateInterconnect

module Network.AWS.DirectConnect.CreatePrivateVirtualInterface

module Network.AWS.DirectConnect.CreatePublicVirtualInterface

module Network.AWS.DirectConnect.DeleteConnection

module Network.AWS.DirectConnect.DeleteInterconnect

module Network.AWS.DirectConnect.DeleteVirtualInterface

module Network.AWS.DirectConnect.DescribeConnections

module Network.AWS.DirectConnect.DescribeConnectionsOnInterconnect

module Network.AWS.DirectConnect.DescribeInterconnects

module Network.AWS.DirectConnect.DescribeLocations

module Network.AWS.DirectConnect.DescribeVirtualGateways

module Network.AWS.DirectConnect.DescribeVirtualInterfaces

module Network.AWS.DirectConnect.Types