Network.AWS.DirectConnect.AllocateConnectionOnInterconnect

Description

Creates a hosted connection on an interconnect.

Allocates a VLAN number and a specified amount of bandwidth for use by a hosted connection on the given interconnect.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_AllocateConnectionOnInterconnect.html

Request

data AllocateConnectionOnInterconnect

Request constructor

allocateConnectionOnInterconnect

Arguments

:: Text
acoiBandwidth

-> Text
acoiConnectionName

-> Text
acoiOwnerAccount

-> Text
acoiInterconnectId

-> Int
acoiVlan

-> AllocateConnectionOnInterconnect

AllocateConnectionOnInterconnect constructor.

The fields accessible through corresponding lenses are:

acoiBandwidth :: Text

acoiConnectionName :: Text

acoiInterconnectId :: Text

acoiOwnerAccount :: Text

acoiVlan :: Int

Request lenses

acoiBandwidth :: Lens' AllocateConnectionOnInterconnect Text

Bandwidth of the connection.

Example: "500Mbps"

Default: None

acoiConnectionName :: Lens' AllocateConnectionOnInterconnect Text

Name of the provisioned connection.

Example: "500M Connection to AWS"

Default: None

acoiInterconnectId :: Lens' AllocateConnectionOnInterconnect Text

ID of the interconnect on which the connection will be provisioned.

Example: dxcon-456abc78

Default: None

acoiOwnerAccount :: Lens' AllocateConnectionOnInterconnect Text

Numeric account Id of the customer for whom the connection will be provisioned.

Example: 123443215678

Default: None

acoiVlan :: Lens' AllocateConnectionOnInterconnect Int

The dedicated VLAN provisioned to the connection.

Example: 101

Default: None

Response

data AllocateConnectionOnInterconnectResponse

Response constructor

allocateConnectionOnInterconnectResponse :: AllocateConnectionOnInterconnectResponse

AllocateConnectionOnInterconnectResponse constructor.

The fields accessible through corresponding lenses are:

acoirBandwidth :: Maybe Text

acoirConnectionId :: Maybe Text

acoirConnectionName :: Maybe Text

acoirConnectionState :: Maybe ConnectionState

acoirLocation :: Maybe Text

acoirOwnerAccount :: Maybe Text

acoirPartnerName :: Maybe Text

acoirRegion :: Maybe Text

acoirVlan :: Maybe Int

Response lenses

acoirBandwidth :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

Bandwidth of the connection.

Example: 1Gbps (for regular connections), or 500Mbps (for hosted connections)

Default: None

acoirConnectionId :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

acoirConnectionName :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

acoirConnectionState :: Lens' AllocateConnectionOnInterconnectResponse (Maybe ConnectionState)

acoirLocation :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

acoirOwnerAccount :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

acoirPartnerName :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

acoirRegion :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

acoirVlan :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Int)