Network.AWS.DirectConnect.AllocatePrivateVirtualInterface

Description

Provisions a private virtual interface to be owned by a different customer. The owner of a connection calls this function to provision a private virtual interface which will be owned by another AWS customer. Virtual interfaces created using this function must be confirmed by the virtual interface owner by calling ConfirmPrivateVirtualInterface. Until this step has been completed, the virtual interface will be in Confirming state, and will not be available for handling traffic.

Request

data AllocatePrivateVirtualInterface

Container for the parameters to the AllocatePrivateVirtualInterface operation.

Request constructor

allocatePrivateVirtualInterface

Arguments

:: Text
apviConnectionId

-> Text
apviOwnerAccount

-> NewPrivateVirtualInterfaceAllocation
apviNewPrivateVirtualInterfaceAllocation

-> AllocatePrivateVirtualInterface

Smart constructor for the minimum required parameters to construct a valid AllocatePrivateVirtualInterface request.

The fields accessible through corresponding lenses are:

ConnectionId :: Text

OwnerAccount :: Text

NewPrivateVirtualInterfaceAllocation :: NewPrivateVirtualInterfaceAllocation

Request lenses

apviConnectionId :: Lens' AllocatePrivateVirtualInterface Text

The connection ID on which the private virtual interface is provisioned. Default: None.

apviOwnerAccount :: Lens' AllocatePrivateVirtualInterface Text

The AWS account that will own the new private virtual interface. Default: None.

apviNewPrivateVirtualInterfaceAllocation :: Lens' AllocatePrivateVirtualInterface NewPrivateVirtualInterfaceAllocation

Detailed information for the private virtual interface to be provisioned. Default: None.

Response

data AllocatePrivateVirtualInterfaceResponse

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

Response constructor

allocatePrivateVirtualInterfaceResponse :: AllocatePrivateVirtualInterfaceResponse

Smart constructor for the minimum required parameters to construct a valid AllocatePrivateVirtualInterfaceResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

OwnerAccount :: Maybe Text

VirtualInterfaceId :: Maybe Text

Location :: Maybe Text

ConnectionId :: Maybe Text

VirtualInterfaceType :: Maybe Text

VirtualInterfaceName :: Maybe Text

Vlan :: Maybe Integer

Asn :: Maybe Integer

AuthKey :: Maybe Text

AmazonAddress :: Maybe Text

CustomerAddress :: Maybe Text

VirtualInterfaceState :: Maybe VirtualInterfaceState

CustomerRouterConfig :: Maybe Text

VirtualGatewayId :: Maybe Text

RouteFilterPrefixes :: [RouteFilterPrefix]

Response lenses

apvirOwnerAccount :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirVirtualInterfaceId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

apvirLocation :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

Where the connection is located. Example: EqSV5 Default: None.

apvirConnectionId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

ID of the connection. Example: dxcon-fg5678gh Default: None.

apvirVirtualInterfaceType :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

apvirVirtualInterfaceName :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

The name of the virtual interface assigned by the customer. Example: "My VPC".

apvirVlan :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Integer)

The VLAN ID. Example: 101.

apvirAsn :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Integer)

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

apvirAuthKey :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

Authentication key for BGP configuration. Example: asdf34example.

apvirAmazonAddress :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

apvirCustomerAddress :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

IP address assigned to the customer interface. Example: 192.168.1.2/30.

apvirVirtualInterfaceState :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

apvirCustomerRouterConfig :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

Information for generating the customer router configuration.

apvirVirtualGatewayId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

apvirRouteFilterPrefixes :: Lens' AllocatePrivateVirtualInterfaceResponse [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).