Network.AWS.DirectConnect.AllocatePublicVirtualInterface

Description

Provisions a public virtual interface to be owned by a different customer.

The owner of a connection calls this function to provision a public virtual interface which will be owned by another AWS customer.

Virtual interfaces created using this function must be confirmed by the virtual interface owner by calling ConfirmPublicVirtualInterface. Until this step has been completed, the virtual interface will be in Confirming state, and will not be available for handling traffic.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_AllocatePublicVirtualInterface.html

Request

data AllocatePublicVirtualInterface

Request constructor

allocatePublicVirtualInterface

Arguments

:: Text
apvi1ConnectionId

-> Text
apvi1OwnerAccount

-> NewPublicVirtualInterfaceAllocation
apvi1NewPublicVirtualInterfaceAllocation

-> AllocatePublicVirtualInterface

AllocatePublicVirtualInterface constructor.

The fields accessible through corresponding lenses are:

apvi1ConnectionId :: Text

apvi1NewPublicVirtualInterfaceAllocation :: NewPublicVirtualInterfaceAllocation

apvi1OwnerAccount :: Text

Request lenses

apvi1ConnectionId :: Lens' AllocatePublicVirtualInterface Text

The connection ID on which the public virtual interface is provisioned.

Default: None

apvi1NewPublicVirtualInterfaceAllocation :: Lens' AllocatePublicVirtualInterface NewPublicVirtualInterfaceAllocation

Detailed information for the public virtual interface to be provisioned.

Default: None

apvi1OwnerAccount :: Lens' AllocatePublicVirtualInterface Text

The AWS account that will own the new public virtual interface.

Default: None

Response

data AllocatePublicVirtualInterfaceResponse

Response constructor

allocatePublicVirtualInterfaceResponse :: AllocatePublicVirtualInterfaceResponse

AllocatePublicVirtualInterfaceResponse constructor.

The fields accessible through corresponding lenses are:

apvir1AmazonAddress :: Maybe Text

apvir1Asn :: Maybe Int

apvir1AuthKey :: Maybe Text

apvir1ConnectionId :: Maybe Text

apvir1CustomerAddress :: Maybe Text

apvir1CustomerRouterConfig :: Maybe Text

apvir1Location :: Maybe Text

apvir1OwnerAccount :: Maybe Text

apvir1RouteFilterPrefixes :: [RouteFilterPrefix]

apvir1VirtualGatewayId :: Maybe Text

apvir1VirtualInterfaceId :: Maybe Text

apvir1VirtualInterfaceName :: Maybe Text

apvir1VirtualInterfaceState :: Maybe VirtualInterfaceState

apvir1VirtualInterfaceType :: Maybe Text

apvir1Vlan :: Maybe Int

Response lenses

apvir1AmazonAddress :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

apvir1Asn :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Int)

apvir1AuthKey :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

apvir1ConnectionId :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

apvir1CustomerAddress :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

apvir1CustomerRouterConfig :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

Information for generating the customer router configuration.

apvir1Location :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

apvir1OwnerAccount :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

apvir1RouteFilterPrefixes :: Lens' AllocatePublicVirtualInterfaceResponse [RouteFilterPrefix]

apvir1VirtualGatewayId :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

apvir1VirtualInterfaceId :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

apvir1VirtualInterfaceName :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

apvir1VirtualInterfaceState :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe VirtualInterfaceState)

apvir1VirtualInterfaceType :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

apvir1Vlan :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Int)