Network.AWS.DirectConnect.AssociateHostedConnection

Description

Associates a hosted connection and its virtual interfaces with a link aggregation group (LAG) or interconnect. If the target interconnect or LAG has an existing hosted connection with a conflicting VLAN number or IP address, the operation fails. This action temporarily interrupts the hosted connection's connectivity to AWS as it is being migrated.

Creating a Request

associateHostedConnection

Arguments

:: Text
assConnectionId

-> Text
assParentConnectionId

-> AssociateHostedConnection

Creates a value of AssociateHostedConnection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

assConnectionId - The ID of the hosted connection. Example: dxcon-abc123 Default: None

assParentConnectionId - The ID of the interconnect or the LAG. Example: dxcon-abc123 or dxlag-abc123 Default: None

data AssociateHostedConnection

Container for the parameters to the AssociateHostedConnection operation.

See: associateHostedConnection smart constructor.

Request Lenses

assConnectionId :: Lens' AssociateHostedConnection Text

The ID of the hosted connection. Example: dxcon-abc123 Default: None

assParentConnectionId :: Lens' AssociateHostedConnection Text

The ID of the interconnect or the LAG. Example: dxcon-abc123 or dxlag-abc123 Default: None

Destructuring the Response

connection :: Connection

Creates a value of Connection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cLagId - Undocumented member.

cVlan - Undocumented member.

cLocation - Undocumented member.

cAwsDevice - The Direct Connection endpoint which the physical connection terminates on.

cConnectionId - Undocumented member.

cLoaIssueTime - The time of the most recent call to DescribeLoa for this connection.

cPartnerName - The name of the AWS Direct Connect service provider associated with the connection.

cConnectionName - Undocumented member.

cBandwidth - Bandwidth of the connection. Example: 1Gbps (for regular connections), or 500Mbps (for hosted connections) Default: None

cOwnerAccount - The AWS account that will own the new connection.

cRegion - Undocumented member.

cConnectionState - Undocumented member.

data Connection

A connection represents the physical network connection between the AWS Direct Connect location and the customer.

See: connection smart constructor.

Response Lenses

cLagId :: Lens' Connection (Maybe Text)

Undocumented member.

cVlan :: Lens' Connection (Maybe Int)

Undocumented member.

cLocation :: Lens' Connection (Maybe Text)

Undocumented member.

cAwsDevice :: Lens' Connection (Maybe Text)

The Direct Connection endpoint which the physical connection terminates on.

cConnectionId :: Lens' Connection (Maybe Text)

Undocumented member.

cLoaIssueTime :: Lens' Connection (Maybe UTCTime)

The time of the most recent call to DescribeLoa for this connection.

cPartnerName :: Lens' Connection (Maybe Text)

The name of the AWS Direct Connect service provider associated with the connection.

cConnectionName :: Lens' Connection (Maybe Text)

Undocumented member.

cBandwidth :: Lens' Connection (Maybe Text)

Bandwidth of the connection. Example: 1Gbps (for regular connections), or 500Mbps (for hosted connections) Default: None

cOwnerAccount :: Lens' Connection (Maybe Text)

The AWS account that will own the new connection.

cRegion :: Lens' Connection (Maybe Text)

Undocumented member.

cConnectionState :: Lens' Connection (Maybe ConnectionState)

Undocumented member.