Network.AWS.DirectConnect.AssociateVirtualInterface

Description

Associates a virtual interface with a specified link aggregation group (LAG) or connection. Connectivity to AWS is temporarily interrupted as the virtual interface is being migrated. If the target connection or LAG has an associated virtual interface with a conflicting VLAN number or a conflicting IP address, the operation fails.

Virtual interfaces associated with a hosted connection cannot be associated with a LAG; hosted connections must be migrated along with their virtual interfaces using AssociateHostedConnection .

Hosted virtual interfaces (an interface for which the owner of the connection is not the owner of physical connection) can only be reassociated by the owner of the physical connection.

Creating a Request

associateVirtualInterface

Arguments

:: Text
aviVirtualInterfaceId

-> Text
aviConnectionId

-> AssociateVirtualInterface

Creates a value of AssociateVirtualInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aviVirtualInterfaceId - The ID of the virtual interface. Example: dxvif-123dfg56 Default: None

aviConnectionId - The ID of the LAG or connection with which to associate the virtual interface. Example: dxlag-abc123 or dxcon-abc123 Default: None

data AssociateVirtualInterface

Container for the parameters to the AssociateVirtualInterface operation.

See: associateVirtualInterface smart constructor.

Request Lenses

aviVirtualInterfaceId :: Lens' AssociateVirtualInterface Text

The ID of the virtual interface. Example: dxvif-123dfg56 Default: None

aviConnectionId :: Lens' AssociateVirtualInterface Text

The ID of the LAG or connection with which to associate the virtual interface. Example: dxlag-abc123 or dxcon-abc123 Default: None

Destructuring the Response

virtualInterface :: VirtualInterface

Creates a value of VirtualInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

viBgpPeers - Undocumented member.

viVirtualGatewayId - Undocumented member.

viRouteFilterPrefixes - Undocumented member.

viCustomerAddress - Undocumented member.

viVlan - Undocumented member.

viLocation - Undocumented member.

viAmazonAddress - Undocumented member.

viAddressFamily - Undocumented member.

viVirtualInterfaceState - Undocumented member.

viConnectionId - Undocumented member.

viVirtualInterfaceType - Undocumented member.

viAsn - Undocumented member.

viAuthKey - Undocumented member.

viCustomerRouterConfig - Information for generating the customer router configuration.

viOwnerAccount - The AWS account that will own the new virtual interface.

viVirtualInterfaceName - Undocumented member.

viVirtualInterfaceId - Undocumented member.

data VirtualInterface

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

See: virtualInterface smart constructor.

Response Lenses

viBgpPeers :: Lens' VirtualInterface [BGPPeer]

Undocumented member.

viVirtualGatewayId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viRouteFilterPrefixes :: Lens' VirtualInterface [RouteFilterPrefix]

Undocumented member.

viCustomerAddress :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVlan :: Lens' VirtualInterface (Maybe Int)

Undocumented member.

viLocation :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAmazonAddress :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAddressFamily :: Lens' VirtualInterface (Maybe AddressFamily)

Undocumented member.

viVirtualInterfaceState :: Lens' VirtualInterface (Maybe VirtualInterfaceState)

Undocumented member.

viConnectionId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVirtualInterfaceType :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAsn :: Lens' VirtualInterface (Maybe Int)

Undocumented member.

viAuthKey :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viCustomerRouterConfig :: Lens' VirtualInterface (Maybe Text)

Information for generating the customer router configuration.

viOwnerAccount :: Lens' VirtualInterface (Maybe Text)

The AWS account that will own the new virtual interface.

viVirtualInterfaceName :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVirtualInterfaceId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.