Network.AWS.DirectConnect.ConfirmConnection

Description

Confirm the creation of a hosted connection on an interconnect. Upon creation, the hosted connection is initially in the Ordering state, and will remain in this state until the owner calls ConfirmConnection to confirm creation of the hosted connection.

Request

data ConfirmConnection

Container for the parameters to the ConfirmConnection operation.

Request constructor

confirmConnection

Arguments

:: Text
ccConnectionId

-> ConfirmConnection

Smart constructor for the minimum required parameters to construct a valid ConfirmConnection request.

The fields accessible through corresponding lenses are:

ConnectionId :: Text

Request lenses

ccConnectionId :: Lens' ConfirmConnection Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

Response

data ConfirmConnectionResponse

The response received when ConfirmConnection is called.

Response constructor

confirmConnectionResponse :: ConfirmConnectionResponse

Smart constructor for the minimum required parameters to construct a valid ConfirmConnectionResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ConnectionState :: Maybe ConnectionState

Response lenses

ccrConnectionState :: Lens' ConfirmConnectionResponse (Maybe ConnectionState)

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.