Network.AWS.DirectConnect.ConfirmPrivateVirtualInterface

Description

Accept ownership of a private virtual interface created by another customer.

After the virtual interface owner calls this function, the virtual interface will be created and attached to the given virtual private gateway, and will be available for handling traffic.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_ConfirmPrivateVirtualInterface.html

Request

data ConfirmPrivateVirtualInterface

Request constructor

confirmPrivateVirtualInterface

Arguments

:: Text
cpviVirtualInterfaceId

-> Text
cpviVirtualGatewayId

-> ConfirmPrivateVirtualInterface

ConfirmPrivateVirtualInterface constructor.

The fields accessible through corresponding lenses are:

cpviVirtualGatewayId :: Text

cpviVirtualInterfaceId :: Text

Request lenses

cpviVirtualGatewayId :: Lens' ConfirmPrivateVirtualInterface Text

ID of the virtual private gateway that will be attached to the virtual interface.

A virtual private gateway can be managed via the Amazon Virtual Private Cloud (VPC) console or the EC2 CreateVpnGateway action.

Default: None

cpviVirtualInterfaceId :: Lens' ConfirmPrivateVirtualInterface Text

Response

data ConfirmPrivateVirtualInterfaceResponse

Response constructor

confirmPrivateVirtualInterfaceResponse :: ConfirmPrivateVirtualInterfaceResponse

ConfirmPrivateVirtualInterfaceResponse constructor.

The fields accessible through corresponding lenses are:

cpvir1VirtualInterfaceState :: Maybe VirtualInterfaceState

Response lenses

cpvir1VirtualInterfaceState :: Lens' ConfirmPrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)