Network.AWS.DirectConnect.ConfirmPublicVirtualInterface

Description

Accept ownership of a public virtual interface created by another customer.

After the virtual interface owner calls this function, the specified virtual interface will be created and made available for handling traffic.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_ConfirmPublicVirtualInterface.html

Request

data ConfirmPublicVirtualInterface

Request constructor

confirmPublicVirtualInterface

Arguments

:: Text
cpvi1VirtualInterfaceId

-> ConfirmPublicVirtualInterface

ConfirmPublicVirtualInterface constructor.

The fields accessible through corresponding lenses are:

cpvi1VirtualInterfaceId :: Text

Request lenses

cpvi1VirtualInterfaceId :: Lens' ConfirmPublicVirtualInterface Text

Response

data ConfirmPublicVirtualInterfaceResponse

Response constructor

confirmPublicVirtualInterfaceResponse :: ConfirmPublicVirtualInterfaceResponse

ConfirmPublicVirtualInterfaceResponse constructor.

The fields accessible through corresponding lenses are:

cpvir3VirtualInterfaceState :: Maybe VirtualInterfaceState

Response lenses

cpvir3VirtualInterfaceState :: Lens' ConfirmPublicVirtualInterfaceResponse (Maybe VirtualInterfaceState)