Network.AWS.DirectConnect.CreateBGPPeer

Description

Creates a new BGP peer on a specified virtual interface. The BGP peer cannot be in the same address family (IPv4/IPv6) of an existing BGP peer on the virtual interface.

You must create a BGP peer for the corresponding address family in order to access AWS resources that also use that address family.

When creating a IPv6 BGP peer, the Amazon address and customer address fields must be left blank. IPv6 addresses are automatically assigned from Amazon's pool of IPv6 addresses; you cannot specify custom IPv6 addresses.

For a public virtual interface, the Autonomous System Number (ASN) must be private or already whitelisted for the virtual interface.

Creating a Request

createBGPPeer :: CreateBGPPeer

Creates a value of CreateBGPPeer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbpNewBGPPeer - Detailed information for the BGP peer to be created. Default: None

cbpVirtualInterfaceId - The ID of the virtual interface on which the BGP peer will be provisioned. Example: dxvif-456abc78 Default: None

data CreateBGPPeer

Container for the parameters to the CreateBGPPeer operation.

See: createBGPPeer smart constructor.

Request Lenses

cbpNewBGPPeer :: Lens' CreateBGPPeer (Maybe NewBGPPeer)

Detailed information for the BGP peer to be created. Default: None

cbpVirtualInterfaceId :: Lens' CreateBGPPeer (Maybe Text)

The ID of the virtual interface on which the BGP peer will be provisioned. Example: dxvif-456abc78 Default: None

Destructuring the Response

createBGPPeerResponse

Arguments

:: Int
cbprsResponseStatus

-> CreateBGPPeerResponse

Creates a value of CreateBGPPeerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbprsVirtualInterface - Undocumented member.

cbprsResponseStatus - -- | The response status code.

data CreateBGPPeerResponse

The response received when CreateBGPPeer is called.

See: createBGPPeerResponse smart constructor.

Response Lenses

cbprsVirtualInterface :: Lens' CreateBGPPeerResponse (Maybe VirtualInterface)

Undocumented member.

cbprsResponseStatus :: Lens' CreateBGPPeerResponse Int

- | The response status code.