Network.AWS.DirectConnect.CreateConnection

Description

Creates a new connection between the customer network and a specific AWS Direct Connect location. A connection links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. An AWS Direct Connect location provides access to Amazon Web Services in the region it is associated with. You can establish connections with AWS Direct Connect locations in multiple regions, but a connection in one region does not provide connectivity to other regions.

Request

data CreateConnection

Container for the parameters to the CreateConnection operation.

Request constructor

mkCreateConnection

Arguments

:: Text
cc1Location

-> Text
cc1Bandwidth

-> Text
cc1ConnectionName

-> CreateConnection

Smart constructor for the minimum required parameters to construct a valid CreateConnection request.

The fields accessible through corresponding lenses are:

Location :: Text

Bandwidth :: Text

ConnectionName :: Text

Request lenses

cc1Location :: Lens' CreateConnection Text

Where the connection is located. Example: EqSV5 Default: None.

cc1Bandwidth :: Lens' CreateConnection Text

Bandwidth of the connection. Example: 1Gbps Default: None.

cc1ConnectionName :: Lens' CreateConnection Text

The name of the connection. Example: "1G Connection to AWS" Default: None.

Response

data CreateConnectionResponse

A connection represents the physical network connection between the AWS Direct Connect location and the customer.

Response constructor

mkCreateConnectionResponse :: CreateConnectionResponse

Smart constructor for the minimum required parameters to construct a valid CreateConnectionResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

OwnerAccount :: Maybe Text

ConnectionId :: Maybe Text

ConnectionName :: Maybe Text

ConnectionState :: Maybe ConnectionState

Region :: Maybe Text

Location :: Maybe Text

Bandwidth :: Maybe Text

Vlan :: Maybe Integer

PartnerName :: Maybe Text

Response lenses

ccrrOwnerAccount :: Lens' CreateConnectionResponse (Maybe Text)

ccrrConnectionId :: Lens' CreateConnectionResponse (Maybe Text)

ID of the connection. Example: dxcon-fg5678gh Default: None.

ccrrConnectionName :: Lens' CreateConnectionResponse (Maybe Text)

The name of the connection. Example: "1G Connection to AWS" Default: None.

ccrrConnectionState :: Lens' CreateConnectionResponse (Maybe ConnectionState)

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

ccrrRegion :: Lens' CreateConnectionResponse (Maybe Text)

The AWS region where the connection is located. Example: us-east-1 Default: None.

ccrrLocation :: Lens' CreateConnectionResponse (Maybe Text)

Where the connection is located. Example: EqSV5 Default: None.

ccrrBandwidth :: Lens' CreateConnectionResponse (Maybe Text)

Bandwidth of the connection. Example: 1Gbps Default: None.

ccrrVlan :: Lens' CreateConnectionResponse (Maybe Integer)

The VLAN ID. Example: 101.

ccrrPartnerName :: Lens' CreateConnectionResponse (Maybe Text)