Network.AWS.DirectConnect.CreateInterconnect

Description

Creates a new interconnect between a AWS Direct Connect partner's network and a specific AWS Direct Connect location.

An interconnect is a connection which is capable of hosting other connections. The AWS Direct Connect partner can use an interconnect to provide sub-1Gbps AWS Direct Connect service to tier 2 customers who do not have their own connections. Like a standard connection, an interconnect links the AWS Direct Connect partner's network to an AWS Direct Connect location over a standard 1 Gbps or 10 Gbps Ethernet fiber-optic cable. One end is connected to the partner's router, the other to an AWS Direct Connect router.

For each end customer, the AWS Direct Connect partner provisions a connection on their interconnect by calling AllocateConnectionOnInterconnect. The end customer can then connect to AWS resources by creating a virtual interface on their connection, using the VLAN assigned to them by the AWS Direct Connect partner.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_CreateInterconnect.html

Request

data CreateInterconnect

Request constructor

createInterconnect

Arguments

:: Text
ciInterconnectName

-> Text
ciBandwidth

-> Text
ciLocation

-> CreateInterconnect

CreateInterconnect constructor.

The fields accessible through corresponding lenses are:

ciBandwidth :: Text

ciInterconnectName :: Text

ciLocation :: Text

Request lenses

ciBandwidth :: Lens' CreateInterconnect Text

The port bandwidth

Example: 1Gbps

Default: None

Available values: 1Gbps,10Gbps

ciInterconnectName :: Lens' CreateInterconnect Text

The name of the interconnect.

Example: "1G Interconnect to AWS"

Default: None

ciLocation :: Lens' CreateInterconnect Text

Where the interconnect is located

Example: EqSV5

Default: None

Response

data CreateInterconnectResponse

Response constructor

createInterconnectResponse :: CreateInterconnectResponse

CreateInterconnectResponse constructor.

The fields accessible through corresponding lenses are:

cirBandwidth :: Maybe Text

cirInterconnectId :: Maybe Text

cirInterconnectName :: Maybe Text

cirInterconnectState :: Maybe InterconnectState

cirLocation :: Maybe Text

cirRegion :: Maybe Text

Response lenses

cirBandwidth :: Lens' CreateInterconnectResponse (Maybe Text)

cirInterconnectId :: Lens' CreateInterconnectResponse (Maybe Text)

cirInterconnectName :: Lens' CreateInterconnectResponse (Maybe Text)

cirInterconnectState :: Lens' CreateInterconnectResponse (Maybe InterconnectState)

cirLocation :: Lens' CreateInterconnectResponse (Maybe Text)

cirRegion :: Lens' CreateInterconnectResponse (Maybe Text)