Network.AWS.DirectConnect.CreateLag

Description

Creates a new link aggregation group (LAG) with the specified number of bundled physical connections between the customer network and a specific AWS Direct Connect location. A LAG is a logical interface that uses the Link Aggregation Control Protocol (LACP) to aggregate multiple 1 gigabit or 10 gigabit interfaces, allowing you to treat them as a single interface.

All connections in a LAG must use the same bandwidth (for example, 10 Gbps), and must terminate at the same AWS Direct Connect endpoint.

You can have up to 10 connections per LAG. Regardless of this limit, if you request more connections for the LAG than AWS Direct Connect can allocate on a single endpoint, no LAG is created.

You can specify an existing physical connection or interconnect to include in the LAG (which counts towards the total number of connections). Doing so interrupts the current physical connection or hosted connections, and re-establishes them as a member of the LAG. The LAG will be created on the same AWS Direct Connect endpoint to which the connection terminates. Any virtual interfaces associated with the connection are automatically disassociated and re-associated with the LAG. The connection ID does not change.

If the AWS account used to create a LAG is a registered AWS Direct Connect partner, the LAG is automatically enabled to host sub-connections. For a LAG owned by a partner, any associated virtual interfaces cannot be directly configured.

Creating a Request

createLag

Arguments

:: Int
clNumberOfConnections

-> Text
clLocation

-> Text
clConnectionsBandwidth

-> Text
clLagName

-> CreateLag

Creates a value of CreateLag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clConnectionId - The ID of an existing connection to migrate to the LAG. Default: None

clNumberOfConnections - The number of physical connections initially provisioned and bundled by the LAG. Default: None

clLocation - The AWS Direct Connect location in which the LAG should be allocated. Example: EqSV5 Default: None

clConnectionsBandwidth - The bandwidth of the individual physical connections bundled by the LAG. Default: None Available values: 1Gbps, 10Gbps

clLagName - The name of the LAG. Example: "3x10G LAG to AWS " Default: None

data CreateLag

Container for the parameters to the CreateLag operation.

See: createLag smart constructor.

Request Lenses

clConnectionId :: Lens' CreateLag (Maybe Text)

The ID of an existing connection to migrate to the LAG. Default: None

clNumberOfConnections :: Lens' CreateLag Int

The number of physical connections initially provisioned and bundled by the LAG. Default: None

clLocation :: Lens' CreateLag Text

The AWS Direct Connect location in which the LAG should be allocated. Example: EqSV5 Default: None

clConnectionsBandwidth :: Lens' CreateLag Text

The bandwidth of the individual physical connections bundled by the LAG. Default: None Available values: 1Gbps, 10Gbps

clLagName :: Lens' CreateLag Text

The name of the LAG. Example: "3x10G LAG to AWS " Default: None

Destructuring the Response

lag :: Lag

Creates a value of Lag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lagLagId - Undocumented member.

lagConnectionsBandwidth - The individual bandwidth of the physical connections bundled by the LAG. Available values: 1Gbps, 10Gbps

lagMinimumLinks - The minimum number of physical connections that must be operational for the LAG itself to be operational. If the number of operational connections drops below this setting, the LAG state changes to down . This value can help to ensure that a LAG is not overutilized if a significant number of its bundled connections go down.

lagLagName - The name of the LAG.

lagLocation - Undocumented member.

lagConnections - A list of connections bundled by this LAG.

lagAwsDevice - The AWS Direct Connection endpoint that hosts the LAG.

lagAllowsHostedConnections - Indicates whether the LAG can host other connections.

lagNumberOfConnections - The number of physical connections bundled by the LAG, up to a maximum of 10.

lagLagState - Undocumented member.

lagOwnerAccount - The owner of the LAG.

lagRegion - Undocumented member.

data Lag

Describes a link aggregation group (LAG). A LAG is a connection that uses the Link Aggregation Control Protocol (LACP) to logically aggregate a bundle of physical connections. Like an interconnect, it can host other connections. All connections in a LAG must terminate on the same physical AWS Direct Connect endpoint, and must be the same bandwidth.

See: lag smart constructor.

Response Lenses

lagLagId :: Lens' Lag (Maybe Text)

Undocumented member.

lagConnectionsBandwidth :: Lens' Lag (Maybe Text)

The individual bandwidth of the physical connections bundled by the LAG. Available values: 1Gbps, 10Gbps

lagMinimumLinks :: Lens' Lag (Maybe Int)

The minimum number of physical connections that must be operational for the LAG itself to be operational. If the number of operational connections drops below this setting, the LAG state changes to down . This value can help to ensure that a LAG is not overutilized if a significant number of its bundled connections go down.

lagLagName :: Lens' Lag (Maybe Text)

The name of the LAG.

lagLocation :: Lens' Lag (Maybe Text)

Undocumented member.

lagConnections :: Lens' Lag [Connection]

A list of connections bundled by this LAG.

lagAwsDevice :: Lens' Lag (Maybe Text)

The AWS Direct Connection endpoint that hosts the LAG.

lagAllowsHostedConnections :: Lens' Lag (Maybe Bool)

Indicates whether the LAG can host other connections.

lagNumberOfConnections :: Lens' Lag (Maybe Int)

The number of physical connections bundled by the LAG, up to a maximum of 10.

lagLagState :: Lens' Lag (Maybe LagState)

Undocumented member.

lagOwnerAccount :: Lens' Lag (Maybe Text)

The owner of the LAG.

lagRegion :: Lens' Lag (Maybe Text)

Undocumented member.