Network.AWS.DirectConnect.CreatePublicVirtualInterface

Description

Creates a new public virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A public virtual interface supports sending traffic to public services of AWS such as Amazon Simple Storage Service (Amazon S3).

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_CreatePublicVirtualInterface.html

Request

data CreatePublicVirtualInterface

Request constructor

createPublicVirtualInterface

Arguments

:: Text
cpviConnectionId

-> NewPublicVirtualInterface
cpviNewPublicVirtualInterface

-> CreatePublicVirtualInterface

CreatePublicVirtualInterface constructor.

The fields accessible through corresponding lenses are:

cpviConnectionId :: Text

cpviNewPublicVirtualInterface :: NewPublicVirtualInterface

Request lenses

cpviConnectionId :: Lens' CreatePublicVirtualInterface Text

cpviNewPublicVirtualInterface :: Lens' CreatePublicVirtualInterface NewPublicVirtualInterface

Detailed information for the public virtual interface to be created.

Default: None

Response

data CreatePublicVirtualInterfaceResponse

Response constructor

createPublicVirtualInterfaceResponse :: CreatePublicVirtualInterfaceResponse

CreatePublicVirtualInterfaceResponse constructor.

The fields accessible through corresponding lenses are:

cpvirAmazonAddress :: Maybe Text

cpvirAsn :: Maybe Int

cpvirAuthKey :: Maybe Text

cpvirConnectionId :: Maybe Text

cpvirCustomerAddress :: Maybe Text

cpvirCustomerRouterConfig :: Maybe Text

cpvirLocation :: Maybe Text

cpvirOwnerAccount :: Maybe Text

cpvirRouteFilterPrefixes :: [RouteFilterPrefix]

cpvirVirtualGatewayId :: Maybe Text

cpvirVirtualInterfaceId :: Maybe Text

cpvirVirtualInterfaceName :: Maybe Text

cpvirVirtualInterfaceState :: Maybe VirtualInterfaceState

cpvirVirtualInterfaceType :: Maybe Text

cpvirVlan :: Maybe Int

Response lenses

cpvirAmazonAddress :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

cpvirAsn :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Int)

cpvirAuthKey :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

cpvirConnectionId :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

cpvirCustomerAddress :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

cpvirCustomerRouterConfig :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

Information for generating the customer router configuration.

cpvirLocation :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

cpvirOwnerAccount :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

cpvirRouteFilterPrefixes :: Lens' CreatePublicVirtualInterfaceResponse [RouteFilterPrefix]

cpvirVirtualGatewayId :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

cpvirVirtualInterfaceId :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

cpvirVirtualInterfaceName :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

cpvirVirtualInterfaceState :: Lens' CreatePublicVirtualInterfaceResponse (Maybe VirtualInterfaceState)

cpvirVirtualInterfaceType :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

cpvirVlan :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Int)