Network.AWS.DirectConnect.DeleteConnection

Description

Deletes the connection.

Deleting a connection only stops the AWS Direct Connect port hour and data transfer charges. You need to cancel separately with the providers any services or charges for cross-connects or network circuits that connect you to the AWS Direct Connect location.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_DeleteConnection.html

Request

data DeleteConnection

Request constructor

deleteConnection

Arguments

:: Text
dcConnectionId

-> DeleteConnection

DeleteConnection constructor.

The fields accessible through corresponding lenses are:

dcConnectionId :: Text

Request lenses

dcConnectionId :: Lens' DeleteConnection Text

Response

data DeleteConnectionResponse

Response constructor

deleteConnectionResponse :: DeleteConnectionResponse

DeleteConnectionResponse constructor.

The fields accessible through corresponding lenses are:

dcrBandwidth :: Maybe Text

dcrConnectionId :: Maybe Text

dcrConnectionName :: Maybe Text

dcrConnectionState :: Maybe ConnectionState

dcrLocation :: Maybe Text

dcrOwnerAccount :: Maybe Text

dcrPartnerName :: Maybe Text

dcrRegion :: Maybe Text

dcrVlan :: Maybe Int

Response lenses

dcrBandwidth :: Lens' DeleteConnectionResponse (Maybe Text)

Bandwidth of the connection.

Example: 1Gbps (for regular connections), or 500Mbps (for hosted connections)

Default: None

dcrConnectionId :: Lens' DeleteConnectionResponse (Maybe Text)

dcrConnectionName :: Lens' DeleteConnectionResponse (Maybe Text)

dcrConnectionState :: Lens' DeleteConnectionResponse (Maybe ConnectionState)

dcrLocation :: Lens' DeleteConnectionResponse (Maybe Text)

dcrOwnerAccount :: Lens' DeleteConnectionResponse (Maybe Text)

dcrPartnerName :: Lens' DeleteConnectionResponse (Maybe Text)

dcrRegion :: Lens' DeleteConnectionResponse (Maybe Text)

dcrVlan :: Lens' DeleteConnectionResponse (Maybe Int)