Network.AWS.DirectConnect.DeleteInterconnect

Description

Deletes the specified interconnect.

Request

data DeleteInterconnect

Container for the parameters to the DeleteInterconnect operation.

Request constructor

mkDeleteInterconnect

Arguments

:: Text
diInterconnectId

-> DeleteInterconnect

Smart constructor for the minimum required parameters to construct a valid DeleteInterconnect request.

The fields accessible through corresponding lenses are:

InterconnectId :: Text

Request lenses

diInterconnectId :: Lens' DeleteInterconnect Text

The ID of the interconnect. Example: dxcon-abc123.

Response

data DeleteInterconnectResponse

The response received when DeleteInterconnect is called.

Response constructor

mkDeleteInterconnectResponse :: DeleteInterconnectResponse

Smart constructor for the minimum required parameters to construct a valid DeleteInterconnectResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

InterconnectState :: Maybe InterconnectState

Response lenses

dirInterconnectState :: Lens' DeleteInterconnectResponse (Maybe InterconnectState)

State of the interconnect. Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The interconnect has been approved, and is being initialized. Available: The network link is up, and the interconnect is ready for use. Down: The network link is down. Deleted: The interconnect has been deleted.