Network.AWS.DirectConnect.DeleteLag

Description

Deletes a link aggregation group (LAG). You cannot delete a LAG if it has active virtual interfaces or hosted connections.

Creating a Request

deleteLag

Arguments

:: Text
dLagId

-> DeleteLag

Creates a value of DeleteLag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dLagId - The ID of the LAG to delete. Example: dxlag-abc123 Default: None

data DeleteLag

Container for the parameters to the DeleteLag operation.

See: deleteLag smart constructor.

Request Lenses

dLagId :: Lens' DeleteLag Text

The ID of the LAG to delete. Example: dxlag-abc123 Default: None

Destructuring the Response

lag :: Lag

Creates a value of Lag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lagLagId - Undocumented member.

lagConnectionsBandwidth - The individual bandwidth of the physical connections bundled by the LAG. Available values: 1Gbps, 10Gbps

lagMinimumLinks - The minimum number of physical connections that must be operational for the LAG itself to be operational. If the number of operational connections drops below this setting, the LAG state changes to down . This value can help to ensure that a LAG is not overutilized if a significant number of its bundled connections go down.

lagLagName - The name of the LAG.

lagLocation - Undocumented member.

lagConnections - A list of connections bundled by this LAG.

lagAwsDevice - The AWS Direct Connection endpoint that hosts the LAG.

lagAllowsHostedConnections - Indicates whether the LAG can host other connections.

lagNumberOfConnections - The number of physical connections bundled by the LAG, up to a maximum of 10.

lagLagState - Undocumented member.

lagOwnerAccount - The owner of the LAG.

lagRegion - Undocumented member.

data Lag

Describes a link aggregation group (LAG). A LAG is a connection that uses the Link Aggregation Control Protocol (LACP) to logically aggregate a bundle of physical connections. Like an interconnect, it can host other connections. All connections in a LAG must terminate on the same physical AWS Direct Connect endpoint, and must be the same bandwidth.

See: lag smart constructor.

Response Lenses

lagLagId :: Lens' Lag (Maybe Text)

Undocumented member.

lagConnectionsBandwidth :: Lens' Lag (Maybe Text)

The individual bandwidth of the physical connections bundled by the LAG. Available values: 1Gbps, 10Gbps

lagMinimumLinks :: Lens' Lag (Maybe Int)

The minimum number of physical connections that must be operational for the LAG itself to be operational. If the number of operational connections drops below this setting, the LAG state changes to down . This value can help to ensure that a LAG is not overutilized if a significant number of its bundled connections go down.

lagLagName :: Lens' Lag (Maybe Text)

The name of the LAG.

lagLocation :: Lens' Lag (Maybe Text)

Undocumented member.

lagConnections :: Lens' Lag [Connection]

A list of connections bundled by this LAG.

lagAwsDevice :: Lens' Lag (Maybe Text)

The AWS Direct Connection endpoint that hosts the LAG.

lagAllowsHostedConnections :: Lens' Lag (Maybe Bool)

Indicates whether the LAG can host other connections.

lagNumberOfConnections :: Lens' Lag (Maybe Int)

The number of physical connections bundled by the LAG, up to a maximum of 10.

lagLagState :: Lens' Lag (Maybe LagState)

Undocumented member.

lagOwnerAccount :: Lens' Lag (Maybe Text)

The owner of the LAG.

lagRegion :: Lens' Lag (Maybe Text)

Undocumented member.