Network.AWS.DirectConnect.DescribeConnectionsOnInterconnect

Description

Return a list of connections that have been provisioned on the given interconnect.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_DescribeConnectionsOnInterconnect.html

Request

data DescribeConnectionsOnInterconnect

Request constructor

describeConnectionsOnInterconnect

Arguments

:: Text
dcoiInterconnectId

-> DescribeConnectionsOnInterconnect

DescribeConnectionsOnInterconnect constructor.

The fields accessible through corresponding lenses are:

dcoiInterconnectId :: Text

Request lenses

dcoiInterconnectId :: Lens' DescribeConnectionsOnInterconnect Text

ID of the interconnect on which a list of connection is provisioned.

Example: dxcon-abc123

Default: None

Response

data DescribeConnectionsOnInterconnectResponse

Response constructor

describeConnectionsOnInterconnectResponse :: DescribeConnectionsOnInterconnectResponse

DescribeConnectionsOnInterconnectResponse constructor.

The fields accessible through corresponding lenses are:

dcoirConnections :: [Connection]

Response lenses

dcoirConnections :: Lens' DescribeConnectionsOnInterconnectResponse [Connection]

A list of connections.