Network.AWS.DirectConnect.DescribeInterconnects

Description

Returns a list of interconnects owned by the AWS account.

If an interconnect ID is provided, it will only return this particular interconnect.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_DescribeInterconnects.html

Request

data DescribeInterconnects

Request constructor

describeInterconnects :: DescribeInterconnects

DescribeInterconnects constructor.

The fields accessible through corresponding lenses are:

diInterconnectId :: Maybe Text

Request lenses

diInterconnectId :: Lens' DescribeInterconnects (Maybe Text)

Response

data DescribeInterconnectsResponse

Response constructor

describeInterconnectsResponse :: DescribeInterconnectsResponse

DescribeInterconnectsResponse constructor.

The fields accessible through corresponding lenses are:

dirInterconnects :: [Interconnect]

Response lenses

dirInterconnects :: Lens' DescribeInterconnectsResponse [Interconnect]

A list of interconnects.