Network.AWS.DirectConnect.DescribeLoa

Description

Returns the LOA-CFA for a connection, interconnect, or link aggregation group (LAG).

The Letter of Authorization - Connecting Facility Assignment (LOA-CFA) is a document that is used when establishing your cross connect to AWS at the colocation facility. For more information, see Requesting Cross Connects at AWS Direct Connect Locations in the AWS Direct Connect user guide.

Creating a Request

describeLoa

Arguments

:: Text
dlConnectionId

-> DescribeLoa

Creates a value of DescribeLoa with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlLoaContentType - A standard media type indicating the content type of the LOA-CFA document. Currently, the only supported value is "applicationpdf". Default: applicationpdf

dlProviderName - The name of the service provider who establishes connectivity on your behalf. If you supply this parameter, the LOA-CFA lists the provider name alongside your company name as the requester of the cross connect. Default: None

dlConnectionId - The ID of a connection, LAG, or interconnect for which to get the LOA-CFA information. Example: dxcon-abc123 or dxlag-abc123 Default: None

data DescribeLoa

Container for the parameters to the DescribeLoa operation.

See: describeLoa smart constructor.

Request Lenses

dlLoaContentType :: Lens' DescribeLoa (Maybe LoaContentType)

A standard media type indicating the content type of the LOA-CFA document. Currently, the only supported value is "applicationpdf". Default: applicationpdf

dlProviderName :: Lens' DescribeLoa (Maybe Text)

The name of the service provider who establishes connectivity on your behalf. If you supply this parameter, the LOA-CFA lists the provider name alongside your company name as the requester of the cross connect. Default: None

dlConnectionId :: Lens' DescribeLoa Text

The ID of a connection, LAG, or interconnect for which to get the LOA-CFA information. Example: dxcon-abc123 or dxlag-abc123 Default: None

Destructuring the Response

describeLoaResponse

Arguments

:: Int
dlrsResponseStatus

-> DescribeLoaResponse

Creates a value of DescribeLoaResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlrsLoaContent - Undocumented member.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

dlrsLoaContentType - Undocumented member.

dlrsResponseStatus - -- | The response status code.

data DescribeLoaResponse

A structure containing the Letter of Authorization - Connecting Facility Assignment (LOA-CFA) for a connection.

See: describeLoaResponse smart constructor.

Response Lenses

dlrsLoaContent :: Lens' DescribeLoaResponse (Maybe ByteString)

Undocumented member.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

dlrsLoaContentType :: Lens' DescribeLoaResponse (Maybe LoaContentType)

Undocumented member.

dlrsResponseStatus :: Lens' DescribeLoaResponse Int

- | The response status code.