Network.AWS.DirectConnect.DescribeLocations

Description

Returns the list of AWS Direct Connect locations in the current AWS region. These are the locations that may be selected when calling CreateConnection or CreateInterconnect.

Creating a Request

describeLocations :: DescribeLocations

Creates a value of DescribeLocations with the minimum fields required to make a request.

data DescribeLocations

See: describeLocations smart constructor.

Destructuring the Response

describeLocationsResponse

Arguments

:: Int
dlsrsResponseStatus

-> DescribeLocationsResponse

Creates a value of DescribeLocationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlsrsLocations - A list of colocation hubs where network providers have equipment. Most regions have multiple locations available.

dlsrsResponseStatus - -- | The response status code.

data DescribeLocationsResponse

A location is a network facility where AWS Direct Connect routers are available to be connected. Generally, these are colocation hubs where many network providers have equipment, and where cross connects can be delivered. Locations include a name and facility code, and must be provided when creating a connection.

See: describeLocationsResponse smart constructor.

Response Lenses

dlsrsLocations :: Lens' DescribeLocationsResponse [Location]

A list of colocation hubs where network providers have equipment. Most regions have multiple locations available.

dlsrsResponseStatus :: Lens' DescribeLocationsResponse Int

- | The response status code.