Network.AWS.DirectConnect.DescribeVirtualGateways

Description

Returns a list of virtual private gateways owned by the AWS account.

You can create one or more AWS Direct Connect private virtual interfaces linking to a virtual private gateway. A virtual private gateway can be managed via Amazon Virtual Private Cloud (VPC) console or the EC2CreateVpnGateway action.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_DescribeVirtualGateways.html

Request

data DescribeVirtualGateways

Request constructor

describeVirtualGateways :: DescribeVirtualGateways

DescribeVirtualGateways constructor.

Response

data DescribeVirtualGatewaysResponse

Response constructor

describeVirtualGatewaysResponse :: DescribeVirtualGatewaysResponse

DescribeVirtualGatewaysResponse constructor.

The fields accessible through corresponding lenses are:

dvgrVirtualGateways :: [VirtualGateway]

Response lenses

dvgrVirtualGateways :: Lens' DescribeVirtualGatewaysResponse [VirtualGateway]

A list of virtual private gateways.