Network.AWS.DirectConnect.DescribeVirtualInterfaces

Description

Displays all virtual interfaces for an AWS account. Virtual interfaces deleted fewer than 15 minutes before DescribeVirtualInterfaces is called are also returned. If a connection ID is included then only virtual interfaces associated with this connection will be returned. If a virtual interface ID is included then only a single virtual interface will be returned.

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

If a connection ID is provided, only virtual interfaces provisioned on the specified connection will be returned. If a virtual interface ID is provided, only this particular virtual interface will be returned.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_DescribeVirtualInterfaces.html

Request

data DescribeVirtualInterfaces

Request constructor

describeVirtualInterfaces :: DescribeVirtualInterfaces

DescribeVirtualInterfaces constructor.

The fields accessible through corresponding lenses are:

dviConnectionId :: Maybe Text

dviVirtualInterfaceId :: Maybe Text

Request lenses

dviConnectionId :: Lens' DescribeVirtualInterfaces (Maybe Text)

dviVirtualInterfaceId :: Lens' DescribeVirtualInterfaces (Maybe Text)

Response

data DescribeVirtualInterfacesResponse

Response constructor

describeVirtualInterfacesResponse :: DescribeVirtualInterfacesResponse

DescribeVirtualInterfacesResponse constructor.

The fields accessible through corresponding lenses are:

dvirVirtualInterfaces :: [VirtualInterface]

Response lenses

dvirVirtualInterfaces :: Lens' DescribeVirtualInterfacesResponse [VirtualInterface]

A list of virtual interfaces.