Network.AWS.DirectConnect.Monadic

Description

AWS Direct Connect links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. With this connection in place, you can create virtual interfaces directly to the AWS cloud and Amazon Virtual Private Cloud, bypassing Internet service providers in your network path.

This module is provided for convenience. It offers an alternative to the common idiom of supplying required fields to an operations's smart constructor, using the operation's lenses to modify additional fields, and then sending or paginating the request.

As an example: using Network.AWS.DirectConnect with the smart constructor and basic lens syntax, before explicitly calling send:

import Control.Monad.Trans.AWS
import Network.AWS.DirectConnect

send $ (mkOperationName w x)
     & onLensField1 .~ y
     & onLensField2 .~ z

Versus using Network.AWS.DirectConnect.Monadic with the State operator variants from Control.Lens.Setter such as .= to modify any additional request parameters before sending:

import Network.AWS.DirectConnect.Monadic

operationName w x $ do
    onLensField1 .= y
    onLensField2 .= z

AllocateConnectionOnInterconnect

Creates a hosted connection on an interconnect. Allocates a VLAN number and a specified amount of bandwidth for use by a hosted connection on the given interconnect.

See: DirectConnect

allocateConnectionOnInterconnect

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
acoiBandwidth

-> Text
acoiConnectionName

-> Text
acoiOwnerAccount

-> Text
acoiInterconnectId

-> Integer
acoiVlan

-> State AllocateConnectionOnInterconnect a

-> m AllocateConnectionOnInterconnectResponse

allocateConnectionOnInterconnectCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
acoiBandwidth

-> Text
acoiConnectionName

-> Text
acoiOwnerAccount

-> Text
acoiInterconnectId

-> Integer
acoiVlan

-> State AllocateConnectionOnInterconnect a

-> m (Either ServiceEr AllocateConnectionOnInterconnectResponse)

AllocatePrivateVirtualInterface

Provisions a private virtual interface to be owned by a different customer. The owner of a connection calls this function to provision a private virtual interface which will be owned by another AWS customer. Virtual interfaces created using this function must be confirmed by the virtual interface owner by calling ConfirmPrivateVirtualInterface. Until this step has been completed, the virtual interface will be in Confirming state, and will not be available for handling traffic.

See: DirectConnect

allocatePrivateVirtualInterface

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
apviConnectionId

-> Text
apviOwnerAccount

-> NewPrivateVirtualInterfaceAllocation
apviNewPrivateVirtualInterfaceAllocation

-> State AllocatePrivateVirtualInterface a

-> m AllocatePrivateVirtualInterfaceResponse

allocatePrivateVirtualInterfaceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
apviConnectionId

-> Text
apviOwnerAccount

-> NewPrivateVirtualInterfaceAllocation
apviNewPrivateVirtualInterfaceAllocation

-> State AllocatePrivateVirtualInterface a

-> m (Either ServiceEr AllocatePrivateVirtualInterfaceResponse)

AllocatePublicVirtualInterface

Provisions a public virtual interface to be owned by a different customer. The owner of a connection calls this function to provision a public virtual interface which will be owned by another AWS customer. Virtual interfaces created using this function must be confirmed by the virtual interface owner by calling ConfirmPublicVirtualInterface. Until this step has been completed, the virtual interface will be in Confirming state, and will not be available for handling traffic.

See: DirectConnect

allocatePublicVirtualInterface

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
apvi1ConnectionId

-> Text
apvi1OwnerAccount

-> NewPublicVirtualInterfaceAllocation
apvi1NewPublicVirtualInterfaceAllocation

-> State AllocatePublicVirtualInterface a

-> m AllocatePublicVirtualInterfaceResponse

allocatePublicVirtualInterfaceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
apvi1ConnectionId

-> Text
apvi1OwnerAccount

-> NewPublicVirtualInterfaceAllocation
apvi1NewPublicVirtualInterfaceAllocation

-> State AllocatePublicVirtualInterface a

-> m (Either ServiceEr AllocatePublicVirtualInterfaceResponse)

ConfirmConnection

Confirm the creation of a hosted connection on an interconnect. Upon creation, the hosted connection is initially in the Ordering state, and will remain in this state until the owner calls ConfirmConnection to confirm creation of the hosted connection.

See: DirectConnect

confirmConnection

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ccConnectionId

-> State ConfirmConnection a

-> m ConfirmConnectionResponse

confirmConnectionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ccConnectionId

-> State ConfirmConnection a

-> m (Either ServiceEr ConfirmConnectionResponse)

ConfirmPrivateVirtualInterface

Accept ownership of a private virtual interface created by another customer. After the virtual interface owner calls this function, the virtual interface will be created and attached to the given virtual private gateway, and will be available for handling traffic.

See: DirectConnect

confirmPrivateVirtualInterface

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cpviVirtualInterfaceId

-> Text
cpviVirtualGatewayId

-> State ConfirmPrivateVirtualInterface a

-> m ConfirmPrivateVirtualInterfaceResponse

confirmPrivateVirtualInterfaceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cpviVirtualInterfaceId

-> Text
cpviVirtualGatewayId

-> State ConfirmPrivateVirtualInterface a

-> m (Either ServiceEr ConfirmPrivateVirtualInterfaceResponse)

ConfirmPublicVirtualInterface

Accept ownership of a public virtual interface created by another customer. After the virtual interface owner calls this function, the specified virtual interface will be created and made available for handling traffic.

See: DirectConnect

confirmPublicVirtualInterface

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cpvi1VirtualInterfaceId

-> State ConfirmPublicVirtualInterface a

-> m ConfirmPublicVirtualInterfaceResponse

confirmPublicVirtualInterfaceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cpvi1VirtualInterfaceId

-> State ConfirmPublicVirtualInterface a

-> m (Either ServiceEr ConfirmPublicVirtualInterfaceResponse)

CreateConnection

Creates a new connection between the customer network and a specific AWS Direct Connect location. A connection links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. An AWS Direct Connect location provides access to Amazon Web Services in the region it is associated with. You can establish connections with AWS Direct Connect locations in multiple regions, but a connection in one region does not provide connectivity to other regions.

See: DirectConnect

createConnection

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cc1Location

-> Text
cc1Bandwidth

-> Text
cc1ConnectionName

-> State CreateConnection a

-> m CreateConnectionResponse

createConnectionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cc1Location

-> Text
cc1Bandwidth

-> Text
cc1ConnectionName

-> State CreateConnection a

-> m (Either ServiceEr CreateConnectionResponse)

CreateInterconnect

Creates a new interconnect between a AWS Direct Connect partner's network and a specific AWS Direct Connect location. An interconnect is a connection which is capable of hosting other connections. The AWS Direct Connect partner can use an interconnect to provide sub-1Gbps AWS Direct Connect service to tier 2 customers who do not have their own connections. Like a standard connection, an interconnect links the AWS Direct Connect partner's network to an AWS Direct Connect location over a standard 1 Gbps or 10 Gbps Ethernet fiber-optic cable. One end is connected to the partner's router, the other to an AWS Direct Connect router. For each end customer, the AWS Direct Connect partner provisions a connection on their interconnect by calling AllocateConnectionOnInterconnect. The end customer can then connect to AWS resources by creating a virtual interface on their connection, using the VLAN assigned to them by the AWS Direct Connect partner.

See: DirectConnect

createInterconnect

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ciInterconnectName

-> Text
ciBandwidth

-> Text
ciLocation

-> State CreateInterconnect a

-> m CreateInterconnectResponse

createInterconnectCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ciInterconnectName

-> Text
ciBandwidth

-> Text
ciLocation

-> State CreateInterconnect a

-> m (Either ServiceEr CreateInterconnectResponse)

CreatePrivateVirtualInterface

Creates a new private virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A private virtual interface supports sending traffic to a single virtual private cloud (VPC).

See: DirectConnect

createPrivateVirtualInterface

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cpvi2ConnectionId

-> NewPrivateVirtualInterface
cpvi2NewPrivateVirtualInterface

-> State CreatePrivateVirtualInterface a

-> m CreatePrivateVirtualInterfaceResponse

createPrivateVirtualInterfaceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cpvi2ConnectionId

-> NewPrivateVirtualInterface
cpvi2NewPrivateVirtualInterface

-> State CreatePrivateVirtualInterface a

-> m (Either ServiceEr CreatePrivateVirtualInterfaceResponse)

CreatePublicVirtualInterface

Creates a new public virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A public virtual interface supports sending traffic to public services of AWS such as Amazon Simple Storage Service (Amazon S3).

See: DirectConnect

createPublicVirtualInterface

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cpvi3ConnectionId

-> NewPublicVirtualInterface
cpvi3NewPublicVirtualInterface

-> State CreatePublicVirtualInterface a

-> m CreatePublicVirtualInterfaceResponse

createPublicVirtualInterfaceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cpvi3ConnectionId

-> NewPublicVirtualInterface
cpvi3NewPublicVirtualInterface

-> State CreatePublicVirtualInterface a

-> m (Either ServiceEr CreatePublicVirtualInterfaceResponse)

DeleteConnection

Deletes the connection. Deleting a connection only stops the AWS Direct Connect port hour and data transfer charges. You need to cancel separately with the providers any services or charges for cross-connects or network circuits that connect you to the AWS Direct Connect location.

See: DirectConnect

deleteConnection

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dcConnectionId

-> State DeleteConnection a

-> m DeleteConnectionResponse

deleteConnectionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dcConnectionId

-> State DeleteConnection a

-> m (Either ServiceEr DeleteConnectionResponse)

DeleteInterconnect

Deletes the specified interconnect.

See: DirectConnect

deleteInterconnect

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
diInterconnectId

-> State DeleteInterconnect a

-> m DeleteInterconnectResponse

deleteInterconnectCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
diInterconnectId

-> State DeleteInterconnect a

-> m (Either ServiceEr DeleteInterconnectResponse)

DeleteVirtualInterface

Deletes a virtual interface.

See: DirectConnect

deleteVirtualInterface

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dviVirtualInterfaceId

-> State DeleteVirtualInterface a

-> m DeleteVirtualInterfaceResponse

deleteVirtualInterfaceCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dviVirtualInterfaceId

-> State DeleteVirtualInterface a

-> m (Either ServiceEr DeleteVirtualInterfaceResponse)

DescribeConnections

Displays all connections in this region. If a connection ID is provided, the call returns only that particular connection.

See: DirectConnect

describeConnections :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeConnections a -> m DescribeConnectionsResponse

describeConnectionsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeConnections a -> m (Either ServiceEr DescribeConnectionsResponse)

DescribeConnectionsOnInterconnect

Return a list of connections that have been provisioned on the given interconnect.

See: DirectConnect

describeConnectionsOnInterconnect

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dcoiInterconnectId

-> State DescribeConnectionsOnInterconnect a

-> m DescribeConnectionsOnInterconnectResponse

describeConnectionsOnInterconnectCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dcoiInterconnectId

-> State DescribeConnectionsOnInterconnect a

-> m (Either ServiceEr DescribeConnectionsOnInterconnectResponse)

DescribeInterconnects

Returns a list of interconnects owned by the AWS account. If an interconnect ID is provided, it will only return this particular interconnect.

See: DirectConnect

describeInterconnects :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeInterconnects a -> m DescribeInterconnectsResponse

describeInterconnectsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeInterconnects a -> m (Either ServiceEr DescribeInterconnectsResponse)

DescribeLocations

Returns the list of AWS Direct Connect locations in the current AWS region. These are the locations that may be selected when calling CreateConnection or CreateInterconnect.

See: DirectConnect

describeLocations :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeLocations a -> m DescribeLocationsResponse

describeLocationsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeLocations a -> m (Either ServiceEr DescribeLocationsResponse)

DescribeVirtualGateways

Returns a list of virtual private gateways owned by the AWS account. You can create one or more AWS Direct Connect private virtual interfaces linking to a virtual private gateway. A virtual private gateway can be managed via Amazon Virtual Private Cloud (VPC) console or the EC2 CreateVpnGateway action.

See: DirectConnect

describeVirtualGateways :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeVirtualGateways a -> m DescribeVirtualGatewaysResponse

describeVirtualGatewaysCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeVirtualGateways a -> m (Either ServiceEr DescribeVirtualGatewaysResponse)

DescribeVirtualInterfaces

Displays all virtual interfaces for an AWS account. Virtual interfaces deleted fewer than 15 minutes before DescribeVirtualInterfaces is called are also returned. If a connection ID is included then only virtual interfaces associated with this connection will be returned. If a virtual interface ID is included then only a single virtual interface will be returned. A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer. If a connection ID is provided, only virtual interfaces provisioned on the specified connection will be returned. If a virtual interface ID is provided, only this particular virtual interface will be returned.

See: DirectConnect

describeVirtualInterfaces :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State DescribeVirtualInterfaces a -> m DescribeVirtualInterfacesResponse

describeVirtualInterfacesCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State DescribeVirtualInterfaces a -> m (Either ServiceEr DescribeVirtualInterfacesResponse)

Re-exported

module Network.AWS.DirectConnect

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()