Network.AWS.DirectConnect.V2012_10_25

Description

AWS Direct Connect links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. With this connection in place, you can create virtual interfaces directly to the AWS cloud and Amazon Virtual Private Cloud, bypassing Internet service providers in your network path.

Documentation

data AllocateConnectionOnInterconnect

Constructors

AllocateConnectionOnInterconnect

Fields

_acoirBandwidth :: Text

Bandwidth of the connection. Example: 1Gbps Default: None.

_acoirConnectionName :: Text

Name of the provisioned connection. Example: "500M Connection to AWS" Default: None.

_acoirInterconnectId :: Text

ID of the interconnect on which the connection will be provisioned. Example: dxcon-456abc78 Default: None.

_acoirOwnerAccount :: Text

Numeric account Id of the customer for whom the connection will be provisioned. Example: 123443215678 Default: None.

_acoirVlan :: Integer

The dedicated VLAN provisioned to the connection. Example: 101 Default: None.

acoirVlan :: Lens' AllocateConnectionOnInterconnect Integer

acoirOwnerAccount :: Lens' AllocateConnectionOnInterconnect Text

acoirInterconnectId :: Lens' AllocateConnectionOnInterconnect Text

acoirConnectionName :: Lens' AllocateConnectionOnInterconnect Text

acoirBandwidth :: Lens' AllocateConnectionOnInterconnect Text

data AllocateConnectionOnInterconnectResponse

Constructors

AllocateConnectionOnInterconnectResponse

Fields

_xBandwidth :: Maybe Text

Bandwidth of the connection. Example: 1Gbps Default: None.

_xConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_xConnectionName :: Maybe Text

The name of the connection. Example: "1G Connection to AWS" Default: None.

_xConnectionState :: Maybe ConnectionState

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

_xLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_xOwnerAccount :: Maybe Text

_xPartnerName :: Maybe Text

_xRegion :: Maybe Text

The AWS region where the connection is located. Example: us-east-1 Default: None.

_xVlan :: Maybe Integer

The VLAN ID. Example: 101.

xVlan :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Integer)

xRegion :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

xPartnerName :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

xOwnerAccount :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

xLocation :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

xConnectionState :: Lens' AllocateConnectionOnInterconnectResponse (Maybe ConnectionState)

xConnectionName :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

xConnectionId :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

xBandwidth :: Lens' AllocateConnectionOnInterconnectResponse (Maybe Text)

data AllocatePrivateVirtualInterface

Constructors

AllocatePrivateVirtualInterface

Fields

_apvirConnectionId :: Text

The connection ID on which the private virtual interface is provisioned. Default: None.

_apvirNewPrivateVirtualInterfaceAllocation :: NewPrivateVirtualInterfaceAllocation

Detailed information for the private virtual interface to be provisioned. Default: None.

_apvirOwnerAccount :: Text

The AWS account that will own the new private virtual interface. Default: None.

apvirOwnerAccount :: Lens' AllocatePrivateVirtualInterface Text

apvirNewPrivateVirtualInterfaceAllocation :: Lens' AllocatePrivateVirtualInterface NewPrivateVirtualInterfaceAllocation

apvirConnectionId :: Lens' AllocatePrivateVirtualInterface Text

data AllocatePrivateVirtualInterfaceResponse

Constructors

AllocatePrivateVirtualInterfaceResponse

Fields

_vjAsn :: Maybe Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_vjAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_vjAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_vjConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_vjCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_vjLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_vjOwnerAccount :: Maybe Text

_vjRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_vjCustomerRouterConfig :: Maybe Text

Information for generating the customer router configuration.

_vjVlan :: Maybe Integer

The VLAN ID. Example: 101.

_vjVirtualGatewayId :: Maybe Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_vjVirtualInterfaceId :: Maybe Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

_vjVirtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

_vjVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

_vjVirtualInterfaceType :: Maybe Text

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

vjVlan :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Integer)

vjVirtualInterfaceType :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjVirtualInterfaceState :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

vjVirtualInterfaceName :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjVirtualInterfaceId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjVirtualGatewayId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjRouteFilterPrefixes :: Lens' AllocatePrivateVirtualInterfaceResponse [RouteFilterPrefix]

vjOwnerAccount :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjLocation :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjCustomerRouterConfig :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjCustomerAddress :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjConnectionId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjAuthKey :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjAsn :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Integer)

vjAmazonAddress :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

data AllocatePublicVirtualInterface

Constructors

AllocatePublicVirtualInterface

Fields

_apvisConnectionId :: Text

The connection ID on which the public virtual interface is provisioned. Default: None.

_apvisNewPublicVirtualInterfaceAllocation :: NewPublicVirtualInterfaceAllocation

Detailed information for the public virtual interface to be provisioned. Default: None.

_apvisOwnerAccount :: Text

The AWS account that will own the new public virtual interface. Default: None.

apvisOwnerAccount :: Lens' AllocatePublicVirtualInterface Text

apvisNewPublicVirtualInterfaceAllocation :: Lens' AllocatePublicVirtualInterface NewPublicVirtualInterfaceAllocation

apvisConnectionId :: Lens' AllocatePublicVirtualInterface Text

data AllocatePublicVirtualInterfaceResponse

Constructors

AllocatePublicVirtualInterfaceResponse

Fields

_vnAsn :: Maybe Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_vnAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_vnAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_vnConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_vnCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_vnLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_vnOwnerAccount :: Maybe Text

_vnRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_vnCustomerRouterConfig :: Maybe Text

Information for generating the customer router configuration.

_vnVlan :: Maybe Integer

The VLAN ID. Example: 101.

_vnVirtualGatewayId :: Maybe Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_vnVirtualInterfaceId :: Maybe Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

_vnVirtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

_vnVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

_vnVirtualInterfaceType :: Maybe Text

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

vnVlan :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Integer)

vnVirtualInterfaceType :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

vnVirtualInterfaceState :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe VirtualInterfaceState)

vnVirtualInterfaceName :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

vnVirtualInterfaceId :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

vnVirtualGatewayId :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

vnRouteFilterPrefixes :: Lens' AllocatePublicVirtualInterfaceResponse [RouteFilterPrefix]

vnOwnerAccount :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

vnLocation :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

vnCustomerRouterConfig :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

vnCustomerAddress :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

vnConnectionId :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

vnAuthKey :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

vnAsn :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Integer)

vnAmazonAddress :: Lens' AllocatePublicVirtualInterfaceResponse (Maybe Text)

data ConfirmConnection

Constructors

ConfirmConnection

Fields

_ccsConnectionId :: Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

ccsConnectionId :: Iso' ConfirmConnection Text

data ConfirmConnectionResponse

Constructors

ConfirmConnectionResponse

Fields

_cctConnectionState :: Maybe ConnectionState

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

cctConnectionState :: Iso' ConfirmConnectionResponse (Maybe ConnectionState)

data ConfirmPrivateVirtualInterface

Constructors

ConfirmPrivateVirtualInterface

Fields

_cpvirVirtualGatewayId :: Text

ID of the virtual private gateway that will be attached to the virtual interface. A virtual private gateway can be managed via the Amazon Virtual Private Cloud (VPC) console or the EC2 CreateVpnGateway action. Default: None.

_cpvirVirtualInterfaceId :: Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

cpvirVirtualInterfaceId :: Lens' ConfirmPrivateVirtualInterface Text

cpvirVirtualGatewayId :: Lens' ConfirmPrivateVirtualInterface Text

data ConfirmPrivateVirtualInterfaceResponse

Constructors

ConfirmPrivateVirtualInterfaceResponse

Fields

_cpvisVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

cpvisVirtualInterfaceState :: Iso' ConfirmPrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

data ConfirmPublicVirtualInterface

Constructors

ConfirmPublicVirtualInterface

Fields

_cpviuVirtualInterfaceId :: Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

cpviuVirtualInterfaceId :: Iso' ConfirmPublicVirtualInterface Text

data ConfirmPublicVirtualInterfaceResponse

Constructors

ConfirmPublicVirtualInterfaceResponse

Fields

_cpvivVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

cpvivVirtualInterfaceState :: Iso' ConfirmPublicVirtualInterfaceResponse (Maybe VirtualInterfaceState)

data CreateConnection

Constructors

CreateConnection

Fields

_ccrBandwidth :: Text

Bandwidth of the connection. Example: 1Gbps Default: None.

_ccrConnectionName :: Text

The name of the connection. Example: "1G Connection to AWS" Default: None.

_ccrLocation :: Text

Where the connection is located. Example: EqSV5 Default: None.

ccrLocation :: Lens' CreateConnection Text

ccrConnectionName :: Lens' CreateConnection Text

ccrBandwidth :: Lens' CreateConnection Text

data CreateConnectionResponse

Constructors

CreateConnectionResponse

Fields

_fBandwidth :: Maybe Text

Bandwidth of the connection. Example: 1Gbps Default: None.

_fConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_fConnectionName :: Maybe Text

The name of the connection. Example: "1G Connection to AWS" Default: None.

_fConnectionState :: Maybe ConnectionState

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

_fLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_fOwnerAccount :: Maybe Text

_fPartnerName :: Maybe Text

_fRegion :: Maybe Text

The AWS region where the connection is located. Example: us-east-1 Default: None.

_fVlan :: Maybe Integer

The VLAN ID. Example: 101.

fVlan :: Lens' CreateConnectionResponse (Maybe Integer)

fRegion :: Lens' CreateConnectionResponse (Maybe Text)

fPartnerName :: Lens' CreateConnectionResponse (Maybe Text)

fOwnerAccount :: Lens' CreateConnectionResponse (Maybe Text)

fLocation :: Lens' CreateConnectionResponse (Maybe Text)

fConnectionState :: Lens' CreateConnectionResponse (Maybe ConnectionState)

fConnectionName :: Lens' CreateConnectionResponse (Maybe Text)

fConnectionId :: Lens' CreateConnectionResponse (Maybe Text)

fBandwidth :: Lens' CreateConnectionResponse (Maybe Text)

data CreateInterconnect

Constructors

CreateInterconnect

Fields

_cirBandwidth :: Text

The port bandwidth Example: 1Gbps Default: None Available values: 1Gbps,10Gbps.

_cirInterconnectName :: Text

The name of the interconnect. Example: "1G Interconnect to AWS" Default: None.

_cirLocation :: Text

Where the interconnect is located Example: EqSV5 Default: None.

cirLocation :: Lens' CreateInterconnect Text

cirInterconnectName :: Lens' CreateInterconnect Text

cirBandwidth :: Lens' CreateInterconnect Text

data CreateInterconnectResponse

Constructors

CreateInterconnectResponse

Fields

_iuBandwidth :: Maybe Text

Bandwidth of the connection. Example: 1Gbps Default: None.

_iuInterconnectId :: Maybe Text

The ID of the interconnect. Example: dxcon-abc123.

_iuInterconnectName :: Maybe Text

The name of the interconnect. Example: "1G Interconnect to AWS".

_iuInterconnectState :: Maybe InterconnectState

State of the interconnect. Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The interconnect has been approved, and is being initialized. Available: The network link is up, and the interconnect is ready for use. Down: The network link is down. Deleted: The interconnect has been deleted.

_iuLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_iuRegion :: Maybe Text

The AWS region where the connection is located. Example: us-east-1 Default: None.

iuRegion :: Lens' CreateInterconnectResponse (Maybe Text)

iuLocation :: Lens' CreateInterconnectResponse (Maybe Text)

iuInterconnectState :: Lens' CreateInterconnectResponse (Maybe InterconnectState)

iuInterconnectName :: Lens' CreateInterconnectResponse (Maybe Text)

iuInterconnectId :: Lens' CreateInterconnectResponse (Maybe Text)

iuBandwidth :: Lens' CreateInterconnectResponse (Maybe Text)

data CreatePrivateVirtualInterface

Constructors

CreatePrivateVirtualInterface

Fields

_cpviwConnectionId :: Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_cpviwNewPrivateVirtualInterface :: NewPrivateVirtualInterface

Detailed information for the private virtual interface to be created. Default: None.

cpviwNewPrivateVirtualInterface :: Lens' CreatePrivateVirtualInterface NewPrivateVirtualInterface

cpviwConnectionId :: Lens' CreatePrivateVirtualInterface Text

data CreatePrivateVirtualInterfaceResponse

Constructors

CreatePrivateVirtualInterfaceResponse

Fields

_vmAsn :: Maybe Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_vmAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_vmAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_vmConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_vmCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_vmLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_vmOwnerAccount :: Maybe Text

_vmRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_vmCustomerRouterConfig :: Maybe Text

Information for generating the customer router configuration.

_vmVlan :: Maybe Integer

The VLAN ID. Example: 101.

_vmVirtualGatewayId :: Maybe Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_vmVirtualInterfaceId :: Maybe Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

_vmVirtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

_vmVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

_vmVirtualInterfaceType :: Maybe Text

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

vmVlan :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Integer)

vmVirtualInterfaceType :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

vmVirtualInterfaceState :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

vmVirtualInterfaceName :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

vmVirtualInterfaceId :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

vmVirtualGatewayId :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

vmRouteFilterPrefixes :: Lens' CreatePrivateVirtualInterfaceResponse [RouteFilterPrefix]

vmOwnerAccount :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

vmLocation :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

vmCustomerRouterConfig :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

vmCustomerAddress :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

vmConnectionId :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

vmAuthKey :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

vmAsn :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Integer)

vmAmazonAddress :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

data CreatePublicVirtualInterface

Constructors

CreatePublicVirtualInterface

Fields

_cpvitConnectionId :: Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_cpvitNewPublicVirtualInterface :: NewPublicVirtualInterface

Detailed information for the public virtual interface to be created. Default: None.

cpvitNewPublicVirtualInterface :: Lens' CreatePublicVirtualInterface NewPublicVirtualInterface

cpvitConnectionId :: Lens' CreatePublicVirtualInterface Text

data CreatePublicVirtualInterfaceResponse

Constructors

CreatePublicVirtualInterfaceResponse

Fields

_viAsn :: Maybe Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_viAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_viAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_viConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_viCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_viLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_viOwnerAccount :: Maybe Text

_viRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_viCustomerRouterConfig :: Maybe Text

Information for generating the customer router configuration.

_viVlan :: Maybe Integer

The VLAN ID. Example: 101.

_viVirtualGatewayId :: Maybe Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_viVirtualInterfaceId :: Maybe Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

_viVirtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

_viVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

_viVirtualInterfaceType :: Maybe Text

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

viVlan :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Integer)

viVirtualInterfaceType :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viVirtualInterfaceState :: Lens' CreatePublicVirtualInterfaceResponse (Maybe VirtualInterfaceState)

viVirtualInterfaceName :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viVirtualInterfaceId :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viVirtualGatewayId :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viRouteFilterPrefixes :: Lens' CreatePublicVirtualInterfaceResponse [RouteFilterPrefix]

viOwnerAccount :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viLocation :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viCustomerRouterConfig :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viCustomerAddress :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viConnectionId :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viAuthKey :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viAsn :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Integer)

viAmazonAddress :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

data DeleteConnection

Constructors

DeleteConnection

Fields

_dcrConnectionId :: Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

dcrConnectionId :: Iso' DeleteConnection Text

data DeleteConnectionResponse

Constructors

DeleteConnectionResponse

Fields

_cBandwidth :: Maybe Text

Bandwidth of the connection. Example: 1Gbps Default: None.

_cConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_cConnectionName :: Maybe Text

The name of the connection. Example: "1G Connection to AWS" Default: None.

_cConnectionState :: Maybe ConnectionState

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

_cLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_cOwnerAccount :: Maybe Text

_cPartnerName :: Maybe Text

_cRegion :: Maybe Text

The AWS region where the connection is located. Example: us-east-1 Default: None.

_cVlan :: Maybe Integer

The VLAN ID. Example: 101.

cVlan :: Lens' DeleteConnectionResponse (Maybe Integer)

cRegion :: Lens' DeleteConnectionResponse (Maybe Text)

cPartnerName :: Lens' DeleteConnectionResponse (Maybe Text)

cOwnerAccount :: Lens' DeleteConnectionResponse (Maybe Text)

cLocation :: Lens' DeleteConnectionResponse (Maybe Text)

cConnectionState :: Lens' DeleteConnectionResponse (Maybe ConnectionState)

cConnectionName :: Lens' DeleteConnectionResponse (Maybe Text)

cConnectionId :: Lens' DeleteConnectionResponse (Maybe Text)

cBandwidth :: Lens' DeleteConnectionResponse (Maybe Text)

data DeleteInterconnect

Constructors

DeleteInterconnect

Fields

_disInterconnectId :: Text

The ID of the interconnect. Example: dxcon-abc123.

disInterconnectId :: Iso' DeleteInterconnect Text

data DeleteInterconnectResponse

Constructors

DeleteInterconnectResponse

Fields

_ditInterconnectState :: Maybe InterconnectState

State of the interconnect. Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The interconnect has been approved, and is being initialized. Available: The network link is up, and the interconnect is ready for use. Down: The network link is down. Deleted: The interconnect has been deleted.

ditInterconnectState :: Iso' DeleteInterconnectResponse (Maybe InterconnectState)

data DeleteVirtualInterface

Constructors

DeleteVirtualInterface

Fields

_dvisVirtualInterfaceId :: Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

dvisVirtualInterfaceId :: Iso' DeleteVirtualInterface Text

data DeleteVirtualInterfaceResponse

Constructors

DeleteVirtualInterfaceResponse

Fields

_dvitVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

dvitVirtualInterfaceState :: Iso' DeleteVirtualInterfaceResponse (Maybe VirtualInterfaceState)

describeConnections :: DescribeConnections

Minimum specification for a DescribeConnections request.

data DescribeConnections

Constructors

DescribeConnections

Fields

_dcsConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

dcsConnectionId :: Iso' DescribeConnections (Maybe Text)

data DescribeConnectionsResponse

Constructors

DescribeConnectionsResponse

Fields

_hConnections :: [Connection]

A list of connections.

hConnections :: Iso' DescribeConnectionsResponse [Connection]

data DescribeConnectionsOnInterconnect

Constructors

DescribeConnectionsOnInterconnect

Fields

_dcoirInterconnectId :: Text

ID of the interconnect on which a list of connection is provisioned. Example: dxcon-abc123 Default: None.

dcoirInterconnectId :: Iso' DescribeConnectionsOnInterconnect Text

data DescribeConnectionsOnInterconnectResponse

Constructors

DescribeConnectionsOnInterconnectResponse

Fields

_mConnections :: [Connection]

A list of connections.

mConnections :: Iso' DescribeConnectionsOnInterconnectResponse [Connection]

describeInterconnects :: DescribeInterconnects

Minimum specification for a DescribeInterconnects request.

data DescribeInterconnects

Constructors

DescribeInterconnects

Fields

_dirInterconnectId :: Maybe Text

The ID of the interconnect. Example: dxcon-abc123.

dirInterconnectId :: Iso' DescribeInterconnects (Maybe Text)

data DescribeInterconnectsResponse

Constructors

DescribeInterconnectsResponse

Fields

_iInterconnects :: [Interconnect]

A list of interconnects.

iInterconnects :: Iso' DescribeInterconnectsResponse [Interconnect]

data DescribeLocations

Constructors

DescribeLocations

data DescribeLocationsResponse

Constructors

DescribeLocationsResponse

Fields

_pLocations :: [Location]

pLocations :: Iso' DescribeLocationsResponse [Location]

data DescribeVirtualGateways

Constructors

DescribeVirtualGateways

data DescribeVirtualGatewaysResponse

Constructors

DescribeVirtualGatewaysResponse

Fields

_vgVirtualGateways :: [VirtualGateway]

A list of virtual private gateways.

vgVirtualGateways :: Iso' DescribeVirtualGatewaysResponse [VirtualGateway]

describeVirtualInterfaces :: DescribeVirtualInterfaces

Minimum specification for a DescribeVirtualInterfaces request.

data DescribeVirtualInterfaces

Constructors

DescribeVirtualInterfaces

Fields

_dvirConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_dvirVirtualInterfaceId :: Maybe Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

dvirVirtualInterfaceId :: Lens' DescribeVirtualInterfaces (Maybe Text)

dvirConnectionId :: Lens' DescribeVirtualInterfaces (Maybe Text)

data DescribeVirtualInterfacesResponse

Constructors

DescribeVirtualInterfacesResponse

Fields

_vkVirtualInterfaces :: [VirtualInterface]

A list of virtual interfaces.

vkVirtualInterfaces :: Iso' DescribeVirtualInterfacesResponse [VirtualInterface]

data DirectConnect

Supported version (2012-10-25) of the AWS Direct Connect service.

data ConnectionState

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

Constructors

ConnectionStateAvailable
available

ConnectionStateDeleted
deleted

ConnectionStateDeleting
deleting

ConnectionStateDown
down

ConnectionStateOrdering
ordering

ConnectionStatePending
pending

ConnectionStateRejected
rejected

ConnectionStateRequested
requested

data InterconnectState

State of the interconnect. Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The interconnect has been approved, and is being initialized. Available: The network link is up, and the interconnect is ready for use. Down: The network link is down. Deleted: The interconnect has been deleted.

Constructors

InterconnectStateAvailable
available

InterconnectStateDeleted
deleted

InterconnectStateDeleting
deleting

InterconnectStateDown
down

InterconnectStatePending
pending

InterconnectStateRequested
requested

data VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

Constructors

VirtualInterfaceStateAvailable
available

VirtualInterfaceStateConfirming
confirming

VirtualInterfaceStateDeleted
deleted

VirtualInterfaceStateDeleting
deleting

VirtualInterfaceStatePending
pending

VirtualInterfaceStateRejected
rejected

VirtualInterfaceStateVerifying
verifying

newtype RouteFilterPrefix

A route filter prefix that the customer can advertise through Border Gateway Protocol (BGP) over a public virtual interface.

Constructors

RouteFilterPrefix

Fields

_rfpCidr :: Maybe Text

CIDR notation for the advertised route. Multiple routes are separated by commas. Example: 10.10.10.024,10.10.11.024.

data Connection

A connection represents the physical network connection between the AWS Direct Connect location and the customer.

Constructors

Connection

Fields

_kVlan :: Maybe Integer

The VLAN ID. Example: 101.

_kLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_kConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_kPartnerName :: Maybe Text

_kConnectionName :: Maybe Text

The name of the connection. Example: "1G Connection to AWS" Default: None.

_kBandwidth :: Maybe Text

Bandwidth of the connection. Example: 1Gbps Default: None.

_kOwnerAccount :: Maybe Text

_kRegion :: Maybe Text

The AWS region where the connection is located. Example: us-east-1 Default: None.

_kConnectionState :: Maybe ConnectionState

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

data Interconnect

An interconnect is a connection that can host other connections. Like a standard AWS Direct Connect connection, an interconnect represents the physical connection between an AWS Direct Connect partner's network and a specific Direct Connect location. An AWS Direct Connect partner who owns an interconnect can provision hosted connections on the interconnect for their end customers, thereby providing the end customers with connectivity to AWS services. The resources of the interconnect, including bandwidth and VLAN numbers, are shared by all of the hosted connections on the interconnect, and the owner of the interconnect determines how these resources are assigned.

Constructors

Interconnect

Fields

_jInterconnectId :: Maybe Text

The ID of the interconnect. Example: dxcon-abc123.

_jLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_jInterconnectName :: Maybe Text

The name of the interconnect. Example: "1G Interconnect to AWS".

_jBandwidth :: Maybe Text

Bandwidth of the connection. Example: 1Gbps Default: None.

_jInterconnectState :: Maybe InterconnectState

State of the interconnect. Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The interconnect has been approved, and is being initialized. Available: The network link is up, and the interconnect is ready for use. Down: The network link is down. Deleted: The interconnect has been deleted.

_jRegion :: Maybe Text

The AWS region where the connection is located. Example: us-east-1 Default: None.

data Location

An AWS Direct Connect location where connections and interconnects can be requested.

Constructors

Location

Fields

_qLocationName :: Maybe Text

The name of the AWS Direct Connect location. The name includes the colocation partner name and the physical site of the lit building.

_qLocationCode :: Maybe Text

The code used to indicate the AWS Direct Connect location.

data NewPrivateVirtualInterface

Detailed information for the private virtual interface to be created. Default: None.

Constructors

NewPrivateVirtualInterface

Fields

_npvjVirtualGatewayId :: Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_npvjCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_npvjVlan :: Integer

The VLAN ID. Example: 101.

_npvjAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_npvjAsn :: Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_npvjAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_npvjVirtualInterfaceName :: Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

data NewPrivateVirtualInterfaceAllocation

Detailed information for the private virtual interface to be provisioned. Default: None.

Constructors

NewPrivateVirtualInterfaceAllocation

Fields

_npviaCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_npviaVlan :: Integer

The VLAN ID. Example: 101.

_npviaAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_npviaAsn :: Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_npviaAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_npviaVirtualInterfaceName :: Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

data NewPublicVirtualInterface

Detailed information for the public virtual interface to be created. Default: None.

Constructors

NewPublicVirtualInterface

Fields

_npviRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_npviCustomerAddress :: Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_npviVlan :: Integer

The VLAN ID. Example: 101.

_npviAmazonAddress :: Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_npviAsn :: Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_npviAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_npviVirtualInterfaceName :: Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

data NewPublicVirtualInterfaceAllocation

Detailed information for the public virtual interface to be provisioned. Default: None.

Constructors

NewPublicVirtualInterfaceAllocation

Fields

_npvibRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_npvibCustomerAddress :: Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_npvibVlan :: Integer

The VLAN ID. Example: 101.

_npvibAmazonAddress :: Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_npvibAsn :: Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_npvibAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_npvibVirtualInterfaceName :: Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

data VirtualGateway

You can create one or more AWS Direct Connect private virtual interfaces linking to your virtual private gateway. Virtual private gateways can be managed using the Amazon Virtual Private Cloud (Amazon VPC) console or the Amazon EC2 CreateVpnGateway action.

Constructors

VirtualGateway

Fields

_vhVirtualGatewayId :: Maybe Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_vhVirtualGatewayState :: Maybe Text

State of the virtual private gateway. Pending: This is the initial state after calling CreateVpnGateway. Available: Ready for use by a private virtual interface. Deleting: This is the initial state after calling DeleteVpnGateway. Deleted: In this state, a private virtual interface is unable to send traffic over this gateway.

data VirtualInterface

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

Constructors

VirtualInterface

Fields

_vlVirtualGatewayId :: Maybe Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_vlRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_vlCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_vlVlan :: Maybe Integer

The VLAN ID. Example: 101.

_vlLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_vlAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_vlVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

_vlConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_vlVirtualInterfaceType :: Maybe Text

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

_vlAsn :: Maybe Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_vlAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_vlCustomerRouterConfig :: Maybe Text

Information for generating the customer router configuration.

_vlOwnerAccount :: Maybe Text

_vlVirtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

_vlVirtualInterfaceId :: Maybe Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

rfpCidr :: Iso' RouteFilterPrefix (Maybe Text)

kVlan :: Lens' Connection (Maybe Integer)

kRegion :: Lens' Connection (Maybe Text)

kPartnerName :: Lens' Connection (Maybe Text)

kOwnerAccount :: Lens' Connection (Maybe Text)

kLocation :: Lens' Connection (Maybe Text)

kConnectionState :: Lens' Connection (Maybe ConnectionState)

kConnectionName :: Lens' Connection (Maybe Text)

kConnectionId :: Lens' Connection (Maybe Text)

kBandwidth :: Lens' Connection (Maybe Text)

jRegion :: Lens' Interconnect (Maybe Text)

jLocation :: Lens' Interconnect (Maybe Text)

jInterconnectState :: Lens' Interconnect (Maybe InterconnectState)

jInterconnectName :: Lens' Interconnect (Maybe Text)

jInterconnectId :: Lens' Interconnect (Maybe Text)

jBandwidth :: Lens' Interconnect (Maybe Text)

qLocationName :: Lens' Location (Maybe Text)

qLocationCode :: Lens' Location (Maybe Text)

npvjVlan :: Lens' NewPrivateVirtualInterface Integer

npvjVirtualInterfaceName :: Lens' NewPrivateVirtualInterface Text

npvjVirtualGatewayId :: Lens' NewPrivateVirtualInterface Text

npvjCustomerAddress :: Lens' NewPrivateVirtualInterface (Maybe Text)

npvjAuthKey :: Lens' NewPrivateVirtualInterface (Maybe Text)

npvjAsn :: Lens' NewPrivateVirtualInterface Integer

npvjAmazonAddress :: Lens' NewPrivateVirtualInterface (Maybe Text)

npviaVlan :: Lens' NewPrivateVirtualInterfaceAllocation Integer

npviaVirtualInterfaceName :: Lens' NewPrivateVirtualInterfaceAllocation Text

npviaCustomerAddress :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

npviaAuthKey :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

npviaAsn :: Lens' NewPrivateVirtualInterfaceAllocation Integer

npviaAmazonAddress :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

npviVlan :: Lens' NewPublicVirtualInterface Integer

npviVirtualInterfaceName :: Lens' NewPublicVirtualInterface Text

npviRouteFilterPrefixes :: Lens' NewPublicVirtualInterface [RouteFilterPrefix]

npviCustomerAddress :: Lens' NewPublicVirtualInterface Text

npviAuthKey :: Lens' NewPublicVirtualInterface (Maybe Text)

npviAsn :: Lens' NewPublicVirtualInterface Integer

npviAmazonAddress :: Lens' NewPublicVirtualInterface Text

npvibVlan :: Lens' NewPublicVirtualInterfaceAllocation Integer

npvibVirtualInterfaceName :: Lens' NewPublicVirtualInterfaceAllocation Text

npvibRouteFilterPrefixes :: Lens' NewPublicVirtualInterfaceAllocation [RouteFilterPrefix]

npvibCustomerAddress :: Lens' NewPublicVirtualInterfaceAllocation Text

npvibAuthKey :: Lens' NewPublicVirtualInterfaceAllocation (Maybe Text)

npvibAsn :: Lens' NewPublicVirtualInterfaceAllocation Integer

npvibAmazonAddress :: Lens' NewPublicVirtualInterfaceAllocation Text

vhVirtualGatewayState :: Lens' VirtualGateway (Maybe Text)

vhVirtualGatewayId :: Lens' VirtualGateway (Maybe Text)

vlVlan :: Lens' VirtualInterface (Maybe Integer)

vlVirtualInterfaceType :: Lens' VirtualInterface (Maybe Text)

vlVirtualInterfaceState :: Lens' VirtualInterface (Maybe VirtualInterfaceState)

vlVirtualInterfaceName :: Lens' VirtualInterface (Maybe Text)

vlVirtualInterfaceId :: Lens' VirtualInterface (Maybe Text)

vlVirtualGatewayId :: Lens' VirtualInterface (Maybe Text)

vlRouteFilterPrefixes :: Lens' VirtualInterface [RouteFilterPrefix]

vlOwnerAccount :: Lens' VirtualInterface (Maybe Text)

vlLocation :: Lens' VirtualInterface (Maybe Text)

vlCustomerRouterConfig :: Lens' VirtualInterface (Maybe Text)

vlCustomerAddress :: Lens' VirtualInterface (Maybe Text)

vlConnectionId :: Lens' VirtualInterface (Maybe Text)

vlAuthKey :: Lens' VirtualInterface (Maybe Text)

vlAsn :: Lens' VirtualInterface (Maybe Integer)

vlAmazonAddress :: Lens' VirtualInterface (Maybe Text)