Network.AWS.DirectConnect.V2012_10_25.AllocatePrivateVirtualInterface

Description

Provisions a private virtual interface to be owned by a different customer. The owner of a connection calls this function to provision a private virtual interface which will be owned by another AWS customer. Virtual interfaces created using this function must be confirmed by the virtual interface owner by calling ConfirmPrivateVirtualInterface. Until this step has been completed, the virtual interface will be in Confirming state, and will not be available for handling traffic.

Documentation

data AllocatePrivateVirtualInterface

Constructors

AllocatePrivateVirtualInterface

Fields

_apvirConnectionId :: Text

The connection ID on which the private virtual interface is provisioned. Default: None.

_apvirNewPrivateVirtualInterfaceAllocation :: NewPrivateVirtualInterfaceAllocation

Detailed information for the private virtual interface to be provisioned. Default: None.

_apvirOwnerAccount :: Text

The AWS account that will own the new private virtual interface. Default: None.

apvirOwnerAccount :: Lens' AllocatePrivateVirtualInterface Text

apvirNewPrivateVirtualInterfaceAllocation :: Lens' AllocatePrivateVirtualInterface NewPrivateVirtualInterfaceAllocation

apvirConnectionId :: Lens' AllocatePrivateVirtualInterface Text

data AllocatePrivateVirtualInterfaceResponse

Constructors

AllocatePrivateVirtualInterfaceResponse

Fields

_vjAsn :: Maybe Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_vjAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_vjAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_vjConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_vjCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_vjLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_vjOwnerAccount :: Maybe Text

_vjRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_vjCustomerRouterConfig :: Maybe Text

Information for generating the customer router configuration.

_vjVlan :: Maybe Integer

The VLAN ID. Example: 101.

_vjVirtualGatewayId :: Maybe Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_vjVirtualInterfaceId :: Maybe Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

_vjVirtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

_vjVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

_vjVirtualInterfaceType :: Maybe Text

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

vjVlan :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Integer)

vjVirtualInterfaceType :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjVirtualInterfaceState :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

vjVirtualInterfaceName :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjVirtualInterfaceId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjVirtualGatewayId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjRouteFilterPrefixes :: Lens' AllocatePrivateVirtualInterfaceResponse [RouteFilterPrefix]

vjOwnerAccount :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjLocation :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjCustomerRouterConfig :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjCustomerAddress :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjConnectionId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjAuthKey :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

vjAsn :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Integer)

vjAmazonAddress :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)