Network.AWS.DirectConnect.V2012_10_25.ConfirmConnection

Description

Confirm the creation of a hosted connection on an interconnect. Upon creation, the hosted connection is initially in the Ordering state, and will remain in this state until the owner calls ConfirmConnection to confirm creation of the hosted connection.

Documentation

data ConfirmConnection

Constructors

ConfirmConnection

Fields

_ccsConnectionId :: Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

ccsConnectionId :: Iso' ConfirmConnection Text

data ConfirmConnectionResponse

Constructors

ConfirmConnectionResponse

Fields

_cctConnectionState :: Maybe ConnectionState

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

cctConnectionState :: Iso' ConfirmConnectionResponse (Maybe ConnectionState)