Network.AWS.DirectConnect.V2012_10_25.CreateInterconnect

Description

Creates a new interconnect between a AWS Direct Connect partner's network and a specific AWS Direct Connect location. An interconnect is a connection which is capable of hosting other connections. The AWS Direct Connect partner can use an interconnect to provide sub-1Gbps AWS Direct Connect service to tier 2 customers who do not have their own connections. Like a standard connection, an interconnect links the AWS Direct Connect partner's network to an AWS Direct Connect location over a standard 1 Gbps or 10 Gbps Ethernet fiber-optic cable. One end is connected to the partner's router, the other to an AWS Direct Connect router. For each end customer, the AWS Direct Connect partner provisions a connection on their interconnect by calling AllocateConnectionOnInterconnect. The end customer can then connect to AWS resources by creating a virtual interface on their connection, using the VLAN assigned to them by the AWS Direct Connect partner.

Request

data CreateInterconnect

Container for the parameters to the CreateInterconnect operation.

Request constructor

mkCreateInterconnect

Arguments

:: Text
ciInterconnectName

-> Text
ciBandwidth

-> Text
ciLocation

-> CreateInterconnect

Smart constructor for the minimum required parameters to construct a valid CreateInterconnect request.

The fields accessible through corresponding lenses are:

InterconnectName :: Text

Bandwidth :: Text

Location :: Text

Request lenses

ciInterconnectName :: Lens' CreateInterconnect Text

The name of the interconnect. Example: "1G Interconnect to AWS" Default: None.

ciBandwidth :: Lens' CreateInterconnect Text

The port bandwidth Example: 1Gbps Default: None Available values: 1Gbps,10Gbps.

ciLocation :: Lens' CreateInterconnect Text

Where the interconnect is located Example: EqSV5 Default: None.

Response

data CreateInterconnectResponse

An interconnect is a connection that can host other connections. Like a standard AWS Direct Connect connection, an interconnect represents the physical connection between an AWS Direct Connect partner's network and a specific Direct Connect location. An AWS Direct Connect partner who owns an interconnect can provision hosted connections on the interconnect for their end customers, thereby providing the end customers with connectivity to AWS services. The resources of the interconnect, including bandwidth and VLAN numbers, are shared by all of the hosted connections on the interconnect, and the owner of the interconnect determines how these resources are assigned.

Response constructor

mkCreateInterconnectResponse :: CreateInterconnectResponse

Smart constructor for the minimum required parameters to construct a valid CreateInterconnectResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

InterconnectId :: Maybe Text

InterconnectName :: Maybe Text

InterconnectState :: Maybe InterconnectState

Region :: Maybe Text

Location :: Maybe Text

Bandwidth :: Maybe Text

Response lenses

cirInterconnectId :: Lens' CreateInterconnectResponse (Maybe Text)

The ID of the interconnect. Example: dxcon-abc123.

cirInterconnectName :: Lens' CreateInterconnectResponse (Maybe Text)

The name of the interconnect. Example: "1G Interconnect to AWS".

cirInterconnectState :: Lens' CreateInterconnectResponse (Maybe InterconnectState)

State of the interconnect. Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The interconnect has been approved, and is being initialized. Available: The network link is up, and the interconnect is ready for use. Down: The network link is down. Deleted: The interconnect has been deleted.

cirRegion :: Lens' CreateInterconnectResponse (Maybe Text)

The AWS region where the connection is located. Example: us-east-1 Default: None.

cirLocation :: Lens' CreateInterconnectResponse (Maybe Text)

Where the connection is located. Example: EqSV5 Default: None.

cirBandwidth :: Lens' CreateInterconnectResponse (Maybe Text)

Bandwidth of the connection. Example: 1Gbps Default: None.