Network.AWS.DirectConnect.V2012_10_25.CreatePublicVirtualInterface

Description

Creates a new public virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A public virtual interface supports sending traffic to public services of AWS such as Amazon Simple Storage Service (Amazon S3).

Documentation

data CreatePublicVirtualInterface

Constructors

CreatePublicVirtualInterface

Fields

_cpvitConnectionId :: Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_cpvitNewPublicVirtualInterface :: NewPublicVirtualInterface

Detailed information for the public virtual interface to be created. Default: None.

cpvitNewPublicVirtualInterface :: Lens' CreatePublicVirtualInterface NewPublicVirtualInterface

cpvitConnectionId :: Lens' CreatePublicVirtualInterface Text

data CreatePublicVirtualInterfaceResponse

Constructors

CreatePublicVirtualInterfaceResponse

Fields

_viAsn :: Maybe Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_viAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_viAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_viConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_viCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_viLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_viOwnerAccount :: Maybe Text

_viRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_viCustomerRouterConfig :: Maybe Text

Information for generating the customer router configuration.

_viVlan :: Maybe Integer

The VLAN ID. Example: 101.

_viVirtualGatewayId :: Maybe Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_viVirtualInterfaceId :: Maybe Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

_viVirtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

_viVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

_viVirtualInterfaceType :: Maybe Text

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

viVlan :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Integer)

viVirtualInterfaceType :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viVirtualInterfaceState :: Lens' CreatePublicVirtualInterfaceResponse (Maybe VirtualInterfaceState)

viVirtualInterfaceName :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viVirtualInterfaceId :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viVirtualGatewayId :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viRouteFilterPrefixes :: Lens' CreatePublicVirtualInterfaceResponse [RouteFilterPrefix]

viOwnerAccount :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viLocation :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viCustomerRouterConfig :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viCustomerAddress :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viConnectionId :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viAuthKey :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)

viAsn :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Integer)

viAmazonAddress :: Lens' CreatePublicVirtualInterfaceResponse (Maybe Text)