Network.AWS.DirectConnect.V2012_10_25.DeleteConnection

Description

Deletes the connection. Deleting a connection only stops the AWS Direct Connect port hour and data transfer charges. You need to cancel separately with the providers any services or charges for cross-connects or network circuits that connect you to the AWS Direct Connect location.

Request

data DeleteConnection

Container for the parameters to the DeleteConnection operation.

Request constructor

mkDeleteConnection

Arguments

:: Text
dcConnectionId

-> DeleteConnection

Smart constructor for the minimum required parameters to construct a valid DeleteConnection request.

Request lenses

dcConnectionId :: Lens' DeleteConnection Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

Response

data DeleteConnectionResponse

A connection represents the physical network connection between the AWS Direct Connect location and the customer.

Response lenses

dcrsOwnerAccount :: Lens' DeleteConnectionResponse (Maybe Text)

dcrsConnectionId :: Lens' DeleteConnectionResponse (Maybe Text)

ID of the connection. Example: dxcon-fg5678gh Default: None.

dcrsConnectionName :: Lens' DeleteConnectionResponse (Maybe Text)

The name of the connection. Example: "1G Connection to AWS" Default: None.

dcrsConnectionState :: Lens' DeleteConnectionResponse (Maybe ConnectionState)

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

dcrsRegion :: Lens' DeleteConnectionResponse (Maybe Text)

The AWS region where the connection is located. Example: us-east-1 Default: None.

dcrsLocation :: Lens' DeleteConnectionResponse (Maybe Text)

Where the connection is located. Example: EqSV5 Default: None.

dcrsBandwidth :: Lens' DeleteConnectionResponse (Maybe Text)

Bandwidth of the connection. Example: 1Gbps Default: None.

dcrsVlan :: Lens' DeleteConnectionResponse (Maybe Integer)

The VLAN ID. Example: 101.

dcrsPartnerName :: Lens' DeleteConnectionResponse (Maybe Text)