Network.AWS.DirectConnect.V2012_10_25.DeleteInterconnect

Description

Deletes the specified interconnect.

Documentation

data DeleteInterconnect

Constructors

DeleteInterconnect

Fields

_disInterconnectId :: Text

The ID of the interconnect. Example: dxcon-abc123.

disInterconnectId :: Iso' DeleteInterconnect Text

data DeleteInterconnectResponse

Constructors

DeleteInterconnectResponse

Fields

_ditInterconnectState :: Maybe InterconnectState

State of the interconnect. Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The interconnect has been approved, and is being initialized. Available: The network link is up, and the interconnect is ready for use. Down: The network link is down. Deleted: The interconnect has been deleted.

ditInterconnectState :: Iso' DeleteInterconnectResponse (Maybe InterconnectState)