Network.AWS.DirectConnect.V2012_10_25.DeleteVirtualInterface

Description

Deletes a virtual interface.

Request

data DeleteVirtualInterface

Container for the parameters to the DeleteVirtualInterface operation.

Request constructor

mkDeleteVirtualInterface

Arguments

:: Text
dviVirtualInterfaceId

-> DeleteVirtualInterface

Smart constructor for the minimum required parameters to construct a valid DeleteVirtualInterface request.

Request lenses

dviVirtualInterfaceId :: Lens' DeleteVirtualInterface Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

Response

data DeleteVirtualInterfaceResponse

The response received when DeleteVirtualInterface is called.

Response lenses

dvirVirtualInterfaceState :: Lens' DeleteVirtualInterfaceResponse (Maybe VirtualInterfaceState)

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.