Network.AWS.DirectConnect.V2012_10_25.DescribeConnectionsOnInterconnect

Description

Return a list of connections that have been provisioned on the given interconnect.

Documentation

data DescribeConnectionsOnInterconnect

Constructors

DescribeConnectionsOnInterconnect

Fields

_dcoirInterconnectId :: Text

ID of the interconnect on which a list of connection is provisioned. Example: dxcon-abc123 Default: None.

dcoirInterconnectId :: Iso' DescribeConnectionsOnInterconnect Text

data DescribeConnectionsOnInterconnectResponse

Constructors

DescribeConnectionsOnInterconnectResponse

Fields

_mConnections :: [Connection]

A list of connections.

mConnections :: Iso' DescribeConnectionsOnInterconnectResponse [Connection]