Network.AWS.DirectConnect.V2012_10_25.DescribeInterconnects

Description

Returns a list of interconnects owned by the AWS account. If an interconnect ID is provided, it will only return this particular interconnect.

Documentation

describeInterconnects :: DescribeInterconnects

Minimum specification for a DescribeInterconnects request.

data DescribeInterconnects

Constructors

DescribeInterconnects

Fields

_dirInterconnectId :: Maybe Text

The ID of the interconnect. Example: dxcon-abc123.

dirInterconnectId :: Iso' DescribeInterconnects (Maybe Text)

data DescribeInterconnectsResponse

Constructors

DescribeInterconnectsResponse

Fields

_iInterconnects :: [Interconnect]

A list of interconnects.

iInterconnects :: Iso' DescribeInterconnectsResponse [Interconnect]