Network.AWS.DirectConnect.V2012_10_25.DescribeVirtualGateways

Description

Returns a list of virtual private gateways owned by the AWS account. You can create one or more AWS Direct Connect private virtual interfaces linking to a virtual private gateway. A virtual private gateway can be managed via Amazon Virtual Private Cloud (VPC) console or the EC2 CreateVpnGateway action.

Documentation

data DescribeVirtualGateways

Constructors

DescribeVirtualGateways

data DescribeVirtualGatewaysResponse

Constructors

DescribeVirtualGatewaysResponse

Fields

_vgVirtualGateways :: [VirtualGateway]

A list of virtual private gateways.

vgVirtualGateways :: Iso' DescribeVirtualGatewaysResponse [VirtualGateway]