Network.AWS.DirectConnect.V2012_10_25.DescribeVirtualInterfaces

Description

Displays all virtual interfaces for an AWS account. Virtual interfaces deleted fewer than 15 minutes before DescribeVirtualInterfaces is called are also returned. If a connection ID is included then only virtual interfaces associated with this connection will be returned. If a virtual interface ID is included then only a single virtual interface will be returned. A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer. If a connection ID is provided, only virtual interfaces provisioned on the specified connection will be returned. If a virtual interface ID is provided, only this particular virtual interface will be returned.

Request

data DescribeVirtualInterfaces

Container for the parameters to the DescribeVirtualInterfaces operation.

Request constructor

mkDescribeVirtualInterfaces :: DescribeVirtualInterfaces

Smart constructor for the minimum required parameters to construct a valid DescribeVirtualInterfaces request.

Request lenses

dvi1ConnectionId :: Lens' DescribeVirtualInterfaces (Maybe Text)

ID of the connection. Example: dxcon-fg5678gh Default: None.

dvi1VirtualInterfaceId :: Lens' DescribeVirtualInterfaces (Maybe Text)

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

Response

data DescribeVirtualInterfacesResponse

A structure containing a list of virtual interfaces.

Response lenses

dvirsrsVirtualInterfaces :: Lens' DescribeVirtualInterfacesResponse [VirtualInterface]

A list of virtual interfaces.