Network.AWS.DirectConnect.V2012_10_25.Types

Description

AWS Direct Connect links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. With this connection in place, you can create virtual interfaces directly to the AWS cloud and Amazon Virtual Private Cloud, bypassing Internet service providers in your network path.

Service

data DirectConnect

Supported version (2012-10-25) of the AWS Direct Connect service.

Errors

data family Er a

ConnectionState

data ConnectionState

Constructors

ConnectionStateAvailable
available

ConnectionStateDeleted
deleted

ConnectionStateDeleting
deleting

ConnectionStateDown
down

ConnectionStateOrdering
ordering

ConnectionStatePending
pending

ConnectionStateRejected
rejected

ConnectionStateRequested
requested

InterconnectState

data InterconnectState

Constructors

InterconnectStateAvailable
available

InterconnectStateDeleted
deleted

InterconnectStateDeleting
deleting

InterconnectStateDown
down

InterconnectStatePending
pending

InterconnectStateRequested
requested

VirtualInterfaceState

data VirtualInterfaceState

Constructors

VirtualInterfaceStateAvailable
available

VirtualInterfaceStateConfirming
confirming

VirtualInterfaceStateDeleted
deleted

VirtualInterfaceStateDeleting
deleting

VirtualInterfaceStatePending
pending

VirtualInterfaceStateRejected
rejected

VirtualInterfaceStateVerifying
verifying

RouteFilterPrefix

data RouteFilterPrefix

A route filter prefix that the customer can advertise through Border Gateway Protocol (BGP) over a public virtual interface.

mkRouteFilterPrefix :: RouteFilterPrefix

Smart constructor for the minimum required fields to construct a valid RouteFilterPrefix data type to populate a request.

rfpCidr :: Lens' RouteFilterPrefix (Maybe Text)

CIDR notation for the advertised route. Multiple routes are separated by commas. Example: 10.10.10.024,10.10.11.024.

Connection

data Connection

A connection represents the physical network connection between the AWS Direct Connect location and the customer.

mkConnection :: Connection

Smart constructor for the minimum required fields to construct a valid Connection data type to populate a request.

cOwnerAccount :: Lens' Connection (Maybe Text)

cConnectionId :: Lens' Connection (Maybe Text)

ID of the connection. Example: dxcon-fg5678gh Default: None.

cConnectionName :: Lens' Connection (Maybe Text)

The name of the connection. Example: "1G Connection to AWS" Default: None.

cConnectionState :: Lens' Connection (Maybe ConnectionState)

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

cRegion :: Lens' Connection (Maybe Text)

The AWS region where the connection is located. Example: us-east-1 Default: None.

cLocation :: Lens' Connection (Maybe Text)

Where the connection is located. Example: EqSV5 Default: None.

cBandwidth :: Lens' Connection (Maybe Text)

Bandwidth of the connection. Example: 1Gbps Default: None.

cVlan :: Lens' Connection (Maybe Integer)

The VLAN ID. Example: 101.

cPartnerName :: Lens' Connection (Maybe Text)

Interconnect

data Interconnect

An interconnect is a connection that can host other connections. Like a standard AWS Direct Connect connection, an interconnect represents the physical connection between an AWS Direct Connect partner's network and a specific Direct Connect location. An AWS Direct Connect partner who owns an interconnect can provision hosted connections on the interconnect for their end customers, thereby providing the end customers with connectivity to AWS services. The resources of the interconnect, including bandwidth and VLAN numbers, are shared by all of the hosted connections on the interconnect, and the owner of the interconnect determines how these resources are assigned.

mkInterconnect :: Interconnect

Smart constructor for the minimum required fields to construct a valid Interconnect data type to populate a request.

iInterconnectId :: Lens' Interconnect (Maybe Text)

The ID of the interconnect. Example: dxcon-abc123.

iInterconnectName :: Lens' Interconnect (Maybe Text)

The name of the interconnect. Example: "1G Interconnect to AWS".

iInterconnectState :: Lens' Interconnect (Maybe InterconnectState)

State of the interconnect. Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The interconnect has been approved, and is being initialized. Available: The network link is up, and the interconnect is ready for use. Down: The network link is down. Deleted: The interconnect has been deleted.

iRegion :: Lens' Interconnect (Maybe Text)

The AWS region where the connection is located. Example: us-east-1 Default: None.

iLocation :: Lens' Interconnect (Maybe Text)

Where the connection is located. Example: EqSV5 Default: None.

iBandwidth :: Lens' Interconnect (Maybe Text)

Bandwidth of the connection. Example: 1Gbps Default: None.

Location

data Location

An AWS Direct Connect location where connections and interconnects can be requested.

mkLocation :: Location

Smart constructor for the minimum required fields to construct a valid Location data type to populate a request.

lLocationCode :: Lens' Location (Maybe Text)

The code used to indicate the AWS Direct Connect location.

lLocationName :: Lens' Location (Maybe Text)

The name of the AWS Direct Connect location. The name includes the colocation partner name and the physical site of the lit building.

NewPrivateVirtualInterface

data NewPrivateVirtualInterface

Detailed information for the private virtual interface to be created. Default: None.

mkNewPrivateVirtualInterface

Arguments

:: Text
npviVirtualInterfaceName

-> Integer
npviVlan

-> Integer
npviAsn

-> Text
npviVirtualGatewayId

-> NewPrivateVirtualInterface

Smart constructor for the minimum required fields to construct a valid NewPrivateVirtualInterface data type to populate a request.

npviVirtualInterfaceName :: Lens' NewPrivateVirtualInterface Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

npviVlan :: Lens' NewPrivateVirtualInterface Integer

The VLAN ID. Example: 101.

npviAsn :: Lens' NewPrivateVirtualInterface Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

npviAuthKey :: Lens' NewPrivateVirtualInterface (Maybe Text)

Authentication key for BGP configuration. Example: asdf34example.

npviAmazonAddress :: Lens' NewPrivateVirtualInterface (Maybe Text)

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

npviCustomerAddress :: Lens' NewPrivateVirtualInterface (Maybe Text)

IP address assigned to the customer interface. Example: 192.168.1.2/30.

npviVirtualGatewayId :: Lens' NewPrivateVirtualInterface Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

NewPrivateVirtualInterfaceAllocation

data NewPrivateVirtualInterfaceAllocation

Detailed information for the private virtual interface to be provisioned. Default: None.

mkNewPrivateVirtualInterfaceAllocation

Arguments

:: Text
npviaVirtualInterfaceName

-> Integer
npviaVlan

-> Integer
npviaAsn

-> NewPrivateVirtualInterfaceAllocation

Smart constructor for the minimum required fields to construct a valid NewPrivateVirtualInterfaceAllocation data type to populate a request.

npviaVirtualInterfaceName :: Lens' NewPrivateVirtualInterfaceAllocation Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

npviaVlan :: Lens' NewPrivateVirtualInterfaceAllocation Integer

The VLAN ID. Example: 101.

npviaAsn :: Lens' NewPrivateVirtualInterfaceAllocation Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

npviaAuthKey :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

Authentication key for BGP configuration. Example: asdf34example.

npviaAmazonAddress :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

npviaCustomerAddress :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

IP address assigned to the customer interface. Example: 192.168.1.2/30.

NewPublicVirtualInterface

data NewPublicVirtualInterface

Detailed information for the public virtual interface to be created. Default: None.

mkNewPublicVirtualInterface

Arguments

:: Text
npvi1VirtualInterfaceName

-> Integer
npvi1Vlan

-> Integer
npvi1Asn

-> Text
npvi1AmazonAddress

-> Text
npvi1CustomerAddress

-> [RouteFilterPrefix]
npvi1RouteFilterPrefixes

-> NewPublicVirtualInterface

Smart constructor for the minimum required fields to construct a valid NewPublicVirtualInterface data type to populate a request.

npvi1VirtualInterfaceName :: Lens' NewPublicVirtualInterface Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

npvi1Vlan :: Lens' NewPublicVirtualInterface Integer

The VLAN ID. Example: 101.

npvi1Asn :: Lens' NewPublicVirtualInterface Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

npvi1AuthKey :: Lens' NewPublicVirtualInterface (Maybe Text)

Authentication key for BGP configuration. Example: asdf34example.

npvi1AmazonAddress :: Lens' NewPublicVirtualInterface Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

npvi1CustomerAddress :: Lens' NewPublicVirtualInterface Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

npvi1RouteFilterPrefixes :: Lens' NewPublicVirtualInterface [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

NewPublicVirtualInterfaceAllocation

data NewPublicVirtualInterfaceAllocation

Detailed information for the public virtual interface to be provisioned. Default: None.

mkNewPublicVirtualInterfaceAllocation

Arguments

:: Text
npvia1VirtualInterfaceName

-> Integer
npvia1Vlan

-> Integer
npvia1Asn

-> Text
npvia1AmazonAddress

-> Text
npvia1CustomerAddress

-> [RouteFilterPrefix]
npvia1RouteFilterPrefixes

-> NewPublicVirtualInterfaceAllocation

Smart constructor for the minimum required fields to construct a valid NewPublicVirtualInterfaceAllocation data type to populate a request.

npvia1VirtualInterfaceName :: Lens' NewPublicVirtualInterfaceAllocation Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

npvia1Vlan :: Lens' NewPublicVirtualInterfaceAllocation Integer

The VLAN ID. Example: 101.

npvia1Asn :: Lens' NewPublicVirtualInterfaceAllocation Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

npvia1AuthKey :: Lens' NewPublicVirtualInterfaceAllocation (Maybe Text)

Authentication key for BGP configuration. Example: asdf34example.

npvia1AmazonAddress :: Lens' NewPublicVirtualInterfaceAllocation Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

npvia1CustomerAddress :: Lens' NewPublicVirtualInterfaceAllocation Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

npvia1RouteFilterPrefixes :: Lens' NewPublicVirtualInterfaceAllocation [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

VirtualGateway

data VirtualGateway

You can create one or more AWS Direct Connect private virtual interfaces linking to your virtual private gateway. Virtual private gateways can be managed using the Amazon Virtual Private Cloud (Amazon VPC) console or the Amazon EC2 CreateVpnGateway action.

mkVirtualGateway :: VirtualGateway

Smart constructor for the minimum required fields to construct a valid VirtualGateway data type to populate a request.

vgVirtualGatewayId :: Lens' VirtualGateway (Maybe Text)

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

vgVirtualGatewayState :: Lens' VirtualGateway (Maybe Text)

State of the virtual private gateway. Pending: This is the initial state after calling CreateVpnGateway. Available: Ready for use by a private virtual interface. Deleting: This is the initial state after calling DeleteVpnGateway. Deleted: In this state, a private virtual interface is unable to send traffic over this gateway.

VirtualInterface

data VirtualInterface

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

mkVirtualInterface :: VirtualInterface

Smart constructor for the minimum required fields to construct a valid VirtualInterface data type to populate a request.

viOwnerAccount :: Lens' VirtualInterface (Maybe Text)

viVirtualInterfaceId :: Lens' VirtualInterface (Maybe Text)

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

viLocation :: Lens' VirtualInterface (Maybe Text)

Where the connection is located. Example: EqSV5 Default: None.

viConnectionId :: Lens' VirtualInterface (Maybe Text)

ID of the connection. Example: dxcon-fg5678gh Default: None.

viVirtualInterfaceType :: Lens' VirtualInterface (Maybe Text)

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

viVirtualInterfaceName :: Lens' VirtualInterface (Maybe Text)

The name of the virtual interface assigned by the customer. Example: "My VPC".

viVlan :: Lens' VirtualInterface (Maybe Integer)

The VLAN ID. Example: 101.

viAsn :: Lens' VirtualInterface (Maybe Integer)

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

viAuthKey :: Lens' VirtualInterface (Maybe Text)

Authentication key for BGP configuration. Example: asdf34example.

viAmazonAddress :: Lens' VirtualInterface (Maybe Text)

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

viCustomerAddress :: Lens' VirtualInterface (Maybe Text)

IP address assigned to the customer interface. Example: 192.168.1.2/30.

viVirtualInterfaceState :: Lens' VirtualInterface (Maybe VirtualInterfaceState)

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

viCustomerRouterConfig :: Lens' VirtualInterface (Maybe Text)

Information for generating the customer router configuration.

viVirtualGatewayId :: Lens' VirtualInterface (Maybe Text)

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

viRouteFilterPrefixes :: Lens' VirtualInterface [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).