Network.AWS.DirectConnect

Description

AWS Direct Connect links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. With this connection in place, you can create virtual interfaces directly to the AWS cloud (for example, to Amazon Elastic Compute Cloud (Amazon EC2) and Amazon Simple Storage Service (Amazon S3)) and to Amazon Virtual Private Cloud (Amazon VPC), bypassing Internet service providers in your network path. An AWS Direct Connect location provides access to AWS in the region it is associated with, as well as access to other US regions. For example, you can provision a single connection to any AWS Direct Connect location in the US and use it to access public AWS services in all US Regions and AWS GovCloud (US).

Service Configuration

directConnect :: Service

API version 2012-10-25 of the Amazon Direct Connect SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by DirectConnect.

DirectConnectClientException

_DirectConnectClientException :: AsError a => Getting (First ServiceError) a ServiceError

The API was called with invalid parameters. The error message will contain additional details about the cause.

DuplicateTagKeysException

_DuplicateTagKeysException :: AsError a => Getting (First ServiceError) a ServiceError

A tag key was specified more than once.

TooManyTagsException

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

You have reached the limit on the number of tags that can be assigned to a Direct Connect resource.

DirectConnectServerException

_DirectConnectServerException :: AsError a => Getting (First ServiceError) a ServiceError

A server-side error occurred during the API call. The error message will contain additional details about the cause.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DescribeInterconnects

module Network.AWS.DirectConnect.DescribeInterconnects

DescribeTags

module Network.AWS.DirectConnect.DescribeTags

DescribeLoa

module Network.AWS.DirectConnect.DescribeLoa

DeleteConnection

module Network.AWS.DirectConnect.DeleteConnection

AssociateConnectionWithLag

module Network.AWS.DirectConnect.AssociateConnectionWithLag

CreateConnection

module Network.AWS.DirectConnect.CreateConnection

AssociateVirtualInterface

module Network.AWS.DirectConnect.AssociateVirtualInterface

DescribeConnections

module Network.AWS.DirectConnect.DescribeConnections

DeleteInterconnect

module Network.AWS.DirectConnect.DeleteInterconnect

ConfirmPrivateVirtualInterface

module Network.AWS.DirectConnect.ConfirmPrivateVirtualInterface

DescribeLocations

module Network.AWS.DirectConnect.DescribeLocations

CreatePublicVirtualInterface

module Network.AWS.DirectConnect.CreatePublicVirtualInterface

AllocatePrivateVirtualInterface

module Network.AWS.DirectConnect.AllocatePrivateVirtualInterface

DescribeLags

module Network.AWS.DirectConnect.DescribeLags

ConfirmConnection

module Network.AWS.DirectConnect.ConfirmConnection

ConfirmPublicVirtualInterface

module Network.AWS.DirectConnect.ConfirmPublicVirtualInterface

DescribeVirtualGateways

module Network.AWS.DirectConnect.DescribeVirtualGateways

DescribeVirtualInterfaces

module Network.AWS.DirectConnect.DescribeVirtualInterfaces

AllocateHostedConnection

module Network.AWS.DirectConnect.AllocateHostedConnection

DeleteVirtualInterface

module Network.AWS.DirectConnect.DeleteVirtualInterface

CreatePrivateVirtualInterface

module Network.AWS.DirectConnect.CreatePrivateVirtualInterface

AllocatePublicVirtualInterface

module Network.AWS.DirectConnect.AllocatePublicVirtualInterface

DisassociateConnectionFromLag

module Network.AWS.DirectConnect.DisassociateConnectionFromLag

TagResource

module Network.AWS.DirectConnect.TagResource

DeleteLag

module Network.AWS.DirectConnect.DeleteLag

UpdateLag

module Network.AWS.DirectConnect.UpdateLag

UntagResource

module Network.AWS.DirectConnect.UntagResource

CreateBGPPeer

module Network.AWS.DirectConnect.CreateBGPPeer

AssociateHostedConnection

module Network.AWS.DirectConnect.AssociateHostedConnection

CreateInterconnect

module Network.AWS.DirectConnect.CreateInterconnect

DeleteBGPPeer

module Network.AWS.DirectConnect.DeleteBGPPeer

CreateLag

module Network.AWS.DirectConnect.CreateLag

DescribeHostedConnections

module Network.AWS.DirectConnect.DescribeHostedConnections

Types

AddressFamily

data AddressFamily

Indicates the address family for the BGP peer.

ipv4 : IPv4 address family

ipv6 : IPv6 address family

Constructors

IPV4

IPV6

BGPPeerState

data BGPPeerState

The state of the BGP peer.

Verifying : The BGP peering addresses or ASN require validation before the BGP peer can be created. This state only applies to BGP peers on a public virtual interface.

Pending : The BGP peer has been created, and is in this state until it is ready to be established.

Available : The BGP peer can be established.

Deleting : The BGP peer is in the process of being deleted.

Deleted : The BGP peer has been deleted and cannot be established.

Constructors

Available

Deleted

Deleting

Pending

Verifying

BGPStatus

data BGPStatus

The Up/Down state of the BGP peer.

Up : The BGP peer is established.

Down : The BGP peer is down.

Constructors

Down

UP

ConnectionState

data ConnectionState

State of the connection.

Ordering : The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order.

Requested : The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

Pending : The connection has been approved, and is being initialized.

Available : The network link is up, and the connection is ready for use.

Down : The network link is down.

Deleting : The connection is in the process of being deleted.

Deleted : The connection has been deleted.

Rejected : A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

Constructors

CSAvailable

CSDeleted

CSDeleting

CSDown

CSOrdering

CSPending

CSRejected

CSRequested

InterconnectState

data InterconnectState

State of the interconnect.

Requested : The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer.

Pending : The interconnect has been approved, and is being initialized.

Available : The network link is up, and the interconnect is ready for use.

Down : The network link is down.

Deleting : The interconnect is in the process of being deleted.

Deleted : The interconnect has been deleted.

Constructors

ISAvailable

ISDeleted

ISDeleting

ISDown

ISPending

ISRequested

LagState

data LagState

The state of the LAG.

Requested : The initial state of a LAG. The LAG stays in the requested state until the Letter of Authorization (LOA) is available.

Pending : The LAG has been approved, and is being initialized.

Available : The network link is established, and the LAG is ready for use.

Down : The network link is down.

Deleting : The LAG is in the process of being deleted.

Deleted : The LAG has been deleted.

Constructors

LSAvailable

LSDeleted

LSDeleting

LSDown

LSPending

LSRequested

LoaContentType

data LoaContentType

A standard media type indicating the content type of the LOA-CFA document. Currently, the only supported value is "application/pdf".

Default: application/pdf

Constructors

ApplicationPdf

VirtualInterfaceState

data VirtualInterfaceState

State of the virtual interface.

Confirming : The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner.

Verifying : This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created.

Pending : A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic.

Available : A virtual interface that is able to forward traffic.

Down : A virtual interface that is BGP down.

Deleting : A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic.

Deleted : A virtual interface that cannot forward traffic.

Rejected : The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

Constructors

VISAvailable

VISConfirming

VISDeleted

VISDeleting

VISDown

VISPending

VISRejected

VISVerifying

BGPPeer

data BGPPeer

A structure containing information about a BGP peer.

See: bgpPeer smart constructor.

bgpPeer :: BGPPeer

Creates a value of BGPPeer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bpCustomerAddress - Undocumented member.

bpAmazonAddress - Undocumented member.

bpAddressFamily - Undocumented member.

bpBgpStatus - Undocumented member.

bpAsn - Undocumented member.

bpAuthKey - Undocumented member.

bpBgpPeerState - Undocumented member.

bpCustomerAddress :: Lens' BGPPeer (Maybe Text)

Undocumented member.

bpAmazonAddress :: Lens' BGPPeer (Maybe Text)

Undocumented member.

bpAddressFamily :: Lens' BGPPeer (Maybe AddressFamily)

Undocumented member.

bpBgpStatus :: Lens' BGPPeer (Maybe BGPStatus)

Undocumented member.

bpAsn :: Lens' BGPPeer (Maybe Int)

Undocumented member.

bpAuthKey :: Lens' BGPPeer (Maybe Text)

Undocumented member.

bpBgpPeerState :: Lens' BGPPeer (Maybe BGPPeerState)

Undocumented member.

Connection

data Connection

A connection represents the physical network connection between the AWS Direct Connect location and the customer.

See: connection smart constructor.

connection :: Connection

Creates a value of Connection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cLagId - Undocumented member.

cVlan - Undocumented member.

cLocation - Undocumented member.

cAwsDevice - The Direct Connection endpoint which the physical connection terminates on.

cConnectionId - Undocumented member.

cLoaIssueTime - The time of the most recent call to DescribeLoa for this connection.

cPartnerName - The name of the AWS Direct Connect service provider associated with the connection.

cConnectionName - Undocumented member.

cBandwidth - Bandwidth of the connection. Example: 1Gbps (for regular connections), or 500Mbps (for hosted connections) Default: None

cOwnerAccount - The AWS account that will own the new connection.

cRegion - Undocumented member.

cConnectionState - Undocumented member.

cLagId :: Lens' Connection (Maybe Text)

Undocumented member.

cVlan :: Lens' Connection (Maybe Int)

Undocumented member.

cLocation :: Lens' Connection (Maybe Text)

Undocumented member.

cAwsDevice :: Lens' Connection (Maybe Text)

The Direct Connection endpoint which the physical connection terminates on.

cConnectionId :: Lens' Connection (Maybe Text)

Undocumented member.

cLoaIssueTime :: Lens' Connection (Maybe UTCTime)

The time of the most recent call to DescribeLoa for this connection.

cPartnerName :: Lens' Connection (Maybe Text)

The name of the AWS Direct Connect service provider associated with the connection.

cConnectionName :: Lens' Connection (Maybe Text)

Undocumented member.

cBandwidth :: Lens' Connection (Maybe Text)

Bandwidth of the connection. Example: 1Gbps (for regular connections), or 500Mbps (for hosted connections) Default: None

cOwnerAccount :: Lens' Connection (Maybe Text)

The AWS account that will own the new connection.

cRegion :: Lens' Connection (Maybe Text)

Undocumented member.

cConnectionState :: Lens' Connection (Maybe ConnectionState)

Undocumented member.

Connections

data Connections

A structure containing a list of connections.

See: connections smart constructor.

connections :: Connections

Creates a value of Connections with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cConnections - A list of connections.

cConnections :: Lens' Connections [Connection]

A list of connections.

Interconnect

data Interconnect

An interconnect is a connection that can host other connections.

Like a standard AWS Direct Connect connection, an interconnect represents the physical connection between an AWS Direct Connect partner's network and a specific Direct Connect location. An AWS Direct Connect partner who owns an interconnect can provision hosted connections on the interconnect for their end customers, thereby providing the end customers with connectivity to AWS services.

The resources of the interconnect, including bandwidth and VLAN numbers, are shared by all of the hosted connections on the interconnect, and the owner of the interconnect determines how these resources are assigned.

See: interconnect smart constructor.

interconnect :: Interconnect

Creates a value of Interconnect with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iLagId - Undocumented member.

iInterconnectId - Undocumented member.

iLocation - Undocumented member.

iInterconnectName - Undocumented member.

iAwsDevice - The Direct Connection endpoint which the physical connection terminates on.

iLoaIssueTime - The time of the most recent call to DescribeInterconnectLoa for this Interconnect.

iBandwidth - Undocumented member.

iInterconnectState - Undocumented member.

iRegion - Undocumented member.

iLagId :: Lens' Interconnect (Maybe Text)

Undocumented member.

iInterconnectId :: Lens' Interconnect (Maybe Text)

Undocumented member.

iLocation :: Lens' Interconnect (Maybe Text)

Undocumented member.

iInterconnectName :: Lens' Interconnect (Maybe Text)

Undocumented member.

iAwsDevice :: Lens' Interconnect (Maybe Text)

The Direct Connection endpoint which the physical connection terminates on.

iLoaIssueTime :: Lens' Interconnect (Maybe UTCTime)

The time of the most recent call to DescribeInterconnectLoa for this Interconnect.

iBandwidth :: Lens' Interconnect (Maybe Text)

Undocumented member.

iInterconnectState :: Lens' Interconnect (Maybe InterconnectState)

Undocumented member.

iRegion :: Lens' Interconnect (Maybe Text)

Undocumented member.

Lag

data Lag

Describes a link aggregation group (LAG). A LAG is a connection that uses the Link Aggregation Control Protocol (LACP) to logically aggregate a bundle of physical connections. Like an interconnect, it can host other connections. All connections in a LAG must terminate on the same physical AWS Direct Connect endpoint, and must be the same bandwidth.

See: lag smart constructor.

lag :: Lag

Creates a value of Lag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lagLagId - Undocumented member.

lagConnectionsBandwidth - The individual bandwidth of the physical connections bundled by the LAG. Available values: 1Gbps, 10Gbps

lagMinimumLinks - The minimum number of physical connections that must be operational for the LAG itself to be operational. If the number of operational connections drops below this setting, the LAG state changes to down . This value can help to ensure that a LAG is not overutilized if a significant number of its bundled connections go down.

lagLagName - The name of the LAG.

lagLocation - Undocumented member.

lagConnections - A list of connections bundled by this LAG.

lagAwsDevice - The AWS Direct Connection endpoint that hosts the LAG.

lagAllowsHostedConnections - Indicates whether the LAG can host other connections.

lagNumberOfConnections - The number of physical connections bundled by the LAG, up to a maximum of 10.

lagLagState - Undocumented member.

lagOwnerAccount - The owner of the LAG.

lagRegion - Undocumented member.

lagLagId :: Lens' Lag (Maybe Text)

Undocumented member.

lagConnectionsBandwidth :: Lens' Lag (Maybe Text)

The individual bandwidth of the physical connections bundled by the LAG. Available values: 1Gbps, 10Gbps

lagMinimumLinks :: Lens' Lag (Maybe Int)

The minimum number of physical connections that must be operational for the LAG itself to be operational. If the number of operational connections drops below this setting, the LAG state changes to down . This value can help to ensure that a LAG is not overutilized if a significant number of its bundled connections go down.

lagLagName :: Lens' Lag (Maybe Text)

The name of the LAG.

lagLocation :: Lens' Lag (Maybe Text)

Undocumented member.

lagConnections :: Lens' Lag [Connection]

A list of connections bundled by this LAG.

lagAwsDevice :: Lens' Lag (Maybe Text)

The AWS Direct Connection endpoint that hosts the LAG.

lagAllowsHostedConnections :: Lens' Lag (Maybe Bool)

Indicates whether the LAG can host other connections.

lagNumberOfConnections :: Lens' Lag (Maybe Int)

The number of physical connections bundled by the LAG, up to a maximum of 10.

lagLagState :: Lens' Lag (Maybe LagState)

Undocumented member.

lagOwnerAccount :: Lens' Lag (Maybe Text)

The owner of the LAG.

lagRegion :: Lens' Lag (Maybe Text)

Undocumented member.

Location

data Location

An AWS Direct Connect location where connections and interconnects can be requested.

See: location smart constructor.

location :: Location

Creates a value of Location with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lLocationName - The name of the AWS Direct Connect location. The name includes the colocation partner name and the physical site of the lit building.

lLocationCode - The code used to indicate the AWS Direct Connect location.

lLocationName :: Lens' Location (Maybe Text)

The name of the AWS Direct Connect location. The name includes the colocation partner name and the physical site of the lit building.

lLocationCode :: Lens' Location (Maybe Text)

The code used to indicate the AWS Direct Connect location.

NewBGPPeer

data NewBGPPeer

A structure containing information about a new BGP peer.

See: newBGPPeer smart constructor.

newBGPPeer :: NewBGPPeer

Creates a value of NewBGPPeer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nbpCustomerAddress - Undocumented member.

nbpAmazonAddress - Undocumented member.

nbpAddressFamily - Undocumented member.

nbpAsn - Undocumented member.

nbpAuthKey - Undocumented member.

nbpCustomerAddress :: Lens' NewBGPPeer (Maybe Text)

Undocumented member.

nbpAmazonAddress :: Lens' NewBGPPeer (Maybe Text)

Undocumented member.

nbpAddressFamily :: Lens' NewBGPPeer (Maybe AddressFamily)

Undocumented member.

nbpAsn :: Lens' NewBGPPeer (Maybe Int)

Undocumented member.

nbpAuthKey :: Lens' NewBGPPeer (Maybe Text)

Undocumented member.

NewPrivateVirtualInterface

data NewPrivateVirtualInterface

A structure containing information about a new private virtual interface.

See: newPrivateVirtualInterface smart constructor.

newPrivateVirtualInterface

Arguments

:: Text
nVirtualInterfaceName

-> Int
nVlan

-> Int
nAsn

-> Text
nVirtualGatewayId

-> NewPrivateVirtualInterface

Creates a value of NewPrivateVirtualInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nCustomerAddress - Undocumented member.

nAmazonAddress - Undocumented member.

nAddressFamily - Undocumented member.

nAuthKey - Undocumented member.

nVirtualInterfaceName - Undocumented member.

nVlan - Undocumented member.

nAsn - Undocumented member.

nVirtualGatewayId - Undocumented member.

nCustomerAddress :: Lens' NewPrivateVirtualInterface (Maybe Text)

Undocumented member.

nAmazonAddress :: Lens' NewPrivateVirtualInterface (Maybe Text)

Undocumented member.

nAddressFamily :: Lens' NewPrivateVirtualInterface (Maybe AddressFamily)

Undocumented member.

nAuthKey :: Lens' NewPrivateVirtualInterface (Maybe Text)

Undocumented member.

nVirtualInterfaceName :: Lens' NewPrivateVirtualInterface Text

Undocumented member.

nVlan :: Lens' NewPrivateVirtualInterface Int

Undocumented member.

nAsn :: Lens' NewPrivateVirtualInterface Int

Undocumented member.

nVirtualGatewayId :: Lens' NewPrivateVirtualInterface Text

Undocumented member.

NewPrivateVirtualInterfaceAllocation

data NewPrivateVirtualInterfaceAllocation

A structure containing information about a private virtual interface that will be provisioned on a connection.

See: newPrivateVirtualInterfaceAllocation smart constructor.

newPrivateVirtualInterfaceAllocation

Arguments

:: Text
npviaVirtualInterfaceName

-> Int
npviaVlan

-> Int
npviaAsn

-> NewPrivateVirtualInterfaceAllocation

Creates a value of NewPrivateVirtualInterfaceAllocation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

npviaCustomerAddress - Undocumented member.

npviaAmazonAddress - Undocumented member.

npviaAddressFamily - Undocumented member.

npviaAuthKey - Undocumented member.

npviaVirtualInterfaceName - Undocumented member.

npviaVlan - Undocumented member.

npviaAsn - Undocumented member.

npviaCustomerAddress :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

Undocumented member.

npviaAmazonAddress :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

Undocumented member.

npviaAddressFamily :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe AddressFamily)

Undocumented member.

npviaAuthKey :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

Undocumented member.

npviaVirtualInterfaceName :: Lens' NewPrivateVirtualInterfaceAllocation Text

Undocumented member.

npviaVlan :: Lens' NewPrivateVirtualInterfaceAllocation Int

Undocumented member.

npviaAsn :: Lens' NewPrivateVirtualInterfaceAllocation Int

Undocumented member.

NewPublicVirtualInterface

data NewPublicVirtualInterface

A structure containing information about a new public virtual interface.

See: newPublicVirtualInterface smart constructor.

newPublicVirtualInterface

Arguments

:: Text
npviVirtualInterfaceName

-> Int
npviVlan

-> Int
npviAsn

-> NewPublicVirtualInterface

Creates a value of NewPublicVirtualInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

npviRouteFilterPrefixes - Undocumented member.

npviCustomerAddress - Undocumented member.

npviAmazonAddress - Undocumented member.

npviAddressFamily - Undocumented member.

npviAuthKey - Undocumented member.

npviVirtualInterfaceName - Undocumented member.

npviVlan - Undocumented member.

npviAsn - Undocumented member.

npviRouteFilterPrefixes :: Lens' NewPublicVirtualInterface [RouteFilterPrefix]

Undocumented member.

npviCustomerAddress :: Lens' NewPublicVirtualInterface (Maybe Text)

Undocumented member.

npviAmazonAddress :: Lens' NewPublicVirtualInterface (Maybe Text)

Undocumented member.

npviAddressFamily :: Lens' NewPublicVirtualInterface (Maybe AddressFamily)

Undocumented member.

npviAuthKey :: Lens' NewPublicVirtualInterface (Maybe Text)

Undocumented member.

npviVirtualInterfaceName :: Lens' NewPublicVirtualInterface Text

Undocumented member.

npviVlan :: Lens' NewPublicVirtualInterface Int

Undocumented member.

npviAsn :: Lens' NewPublicVirtualInterface Int

Undocumented member.

NewPublicVirtualInterfaceAllocation

data NewPublicVirtualInterfaceAllocation

A structure containing information about a public virtual interface that will be provisioned on a connection.

See: newPublicVirtualInterfaceAllocation smart constructor.

newPublicVirtualInterfaceAllocation

Arguments

:: Text
newVirtualInterfaceName

-> Int
newVlan

-> Int
newAsn

-> NewPublicVirtualInterfaceAllocation

Creates a value of NewPublicVirtualInterfaceAllocation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

newRouteFilterPrefixes - Undocumented member.

newCustomerAddress - Undocumented member.

newAmazonAddress - Undocumented member.

newAddressFamily - Undocumented member.

newAuthKey - Undocumented member.

newVirtualInterfaceName - Undocumented member.

newVlan - Undocumented member.

newAsn - Undocumented member.

newRouteFilterPrefixes :: Lens' NewPublicVirtualInterfaceAllocation [RouteFilterPrefix]

Undocumented member.

newCustomerAddress :: Lens' NewPublicVirtualInterfaceAllocation (Maybe Text)

Undocumented member.

newAmazonAddress :: Lens' NewPublicVirtualInterfaceAllocation (Maybe Text)

Undocumented member.

newAddressFamily :: Lens' NewPublicVirtualInterfaceAllocation (Maybe AddressFamily)

Undocumented member.

newAuthKey :: Lens' NewPublicVirtualInterfaceAllocation (Maybe Text)

Undocumented member.

newVirtualInterfaceName :: Lens' NewPublicVirtualInterfaceAllocation Text

Undocumented member.

newVlan :: Lens' NewPublicVirtualInterfaceAllocation Int

Undocumented member.

newAsn :: Lens' NewPublicVirtualInterfaceAllocation Int

Undocumented member.

ResourceTag

data ResourceTag

The tags associated with a Direct Connect resource.

See: resourceTag smart constructor.

resourceTag :: ResourceTag

Creates a value of ResourceTag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtResourceARN - The Amazon Resource Name (ARN) of the Direct Connect resource.

rtTags - The tags.

rtResourceARN :: Lens' ResourceTag (Maybe Text)

The Amazon Resource Name (ARN) of the Direct Connect resource.

rtTags :: Lens' ResourceTag (Maybe (NonEmpty Tag))

The tags.

RouteFilterPrefix

data RouteFilterPrefix

A route filter prefix that the customer can advertise through Border Gateway Protocol (BGP) over a public virtual interface.

See: routeFilterPrefix smart constructor.

routeFilterPrefix :: RouteFilterPrefix

Creates a value of RouteFilterPrefix with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rfpCidr - CIDR notation for the advertised route. Multiple routes are separated by commas. IPv6 CIDRs must be at least a 64 or shorter Example: 10.10.10.024,10.10.11.024,2001:db8::64

rfpCidr :: Lens' RouteFilterPrefix (Maybe Text)

CIDR notation for the advertised route. Multiple routes are separated by commas. IPv6 CIDRs must be at least a 64 or shorter Example: 10.10.10.024,10.10.11.024,2001:db8::64

Tag

data Tag

Information about a tag.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - The value of the tag.

tagKey - The key of the tag.

tagValue :: Lens' Tag (Maybe Text)

The value of the tag.

tagKey :: Lens' Tag Text

The key of the tag.

VirtualGateway

data VirtualGateway

You can create one or more AWS Direct Connect private virtual interfaces linking to your virtual private gateway.

Virtual private gateways can be managed using the Amazon Virtual Private Cloud (Amazon VPC) console or the Amazon EC2 CreateVpnGateway action .

See: virtualGateway smart constructor.

virtualGateway :: VirtualGateway

Creates a value of VirtualGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vgVirtualGatewayId - Undocumented member.

vgVirtualGatewayState - Undocumented member.

vgVirtualGatewayId :: Lens' VirtualGateway (Maybe Text)

Undocumented member.

vgVirtualGatewayState :: Lens' VirtualGateway (Maybe Text)

Undocumented member.

VirtualInterface

data VirtualInterface

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

See: virtualInterface smart constructor.

virtualInterface :: VirtualInterface

Creates a value of VirtualInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

viBgpPeers - Undocumented member.

viVirtualGatewayId - Undocumented member.

viRouteFilterPrefixes - Undocumented member.

viCustomerAddress - Undocumented member.

viVlan - Undocumented member.

viLocation - Undocumented member.

viAmazonAddress - Undocumented member.

viAddressFamily - Undocumented member.

viVirtualInterfaceState - Undocumented member.

viConnectionId - Undocumented member.

viVirtualInterfaceType - Undocumented member.

viAsn - Undocumented member.

viAuthKey - Undocumented member.

viCustomerRouterConfig - Information for generating the customer router configuration.

viOwnerAccount - The AWS account that will own the new virtual interface.

viVirtualInterfaceName - Undocumented member.

viVirtualInterfaceId - Undocumented member.

viBgpPeers :: Lens' VirtualInterface [BGPPeer]

Undocumented member.

viVirtualGatewayId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viRouteFilterPrefixes :: Lens' VirtualInterface [RouteFilterPrefix]

Undocumented member.

viCustomerAddress :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVlan :: Lens' VirtualInterface (Maybe Int)

Undocumented member.

viLocation :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAmazonAddress :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAddressFamily :: Lens' VirtualInterface (Maybe AddressFamily)

Undocumented member.

viVirtualInterfaceState :: Lens' VirtualInterface (Maybe VirtualInterfaceState)

Undocumented member.

viConnectionId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVirtualInterfaceType :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAsn :: Lens' VirtualInterface (Maybe Int)

Undocumented member.

viAuthKey :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viCustomerRouterConfig :: Lens' VirtualInterface (Maybe Text)

Information for generating the customer router configuration.

viOwnerAccount :: Lens' VirtualInterface (Maybe Text)

The AWS account that will own the new virtual interface.

viVirtualInterfaceName :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVirtualInterfaceId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.