Network.AWS.DirectConnect.AllocatePrivateVirtualInterface

Description

Provisions a private virtual interface to be owned by another AWS customer.

Virtual interfaces created using this action must be confirmed by the virtual interface owner by using the ConfirmPrivateVirtualInterface action. Until then, the virtual interface will be in Confirming state, and will not be available for handling traffic.

Creating a Request

allocatePrivateVirtualInterface

Arguments

:: Text
apviConnectionId

-> Text
apviOwnerAccount

-> NewPrivateVirtualInterfaceAllocation
apviNewPrivateVirtualInterfaceAllocation

-> AllocatePrivateVirtualInterface

Creates a value of AllocatePrivateVirtualInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apviConnectionId - The connection ID on which the private virtual interface is provisioned. Default: None

apviOwnerAccount - The AWS account that will own the new private virtual interface. Default: None

apviNewPrivateVirtualInterfaceAllocation - Detailed information for the private virtual interface to be provisioned. Default: None

data AllocatePrivateVirtualInterface

Container for the parameters to the AllocatePrivateVirtualInterface operation.

See: allocatePrivateVirtualInterface smart constructor.

Request Lenses

apviConnectionId :: Lens' AllocatePrivateVirtualInterface Text

The connection ID on which the private virtual interface is provisioned. Default: None

apviOwnerAccount :: Lens' AllocatePrivateVirtualInterface Text

The AWS account that will own the new private virtual interface. Default: None

apviNewPrivateVirtualInterfaceAllocation :: Lens' AllocatePrivateVirtualInterface NewPrivateVirtualInterfaceAllocation

Detailed information for the private virtual interface to be provisioned. Default: None

Destructuring the Response

virtualInterface :: VirtualInterface

Creates a value of VirtualInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

viBgpPeers - Undocumented member.

viVirtualGatewayId - Undocumented member.

viRouteFilterPrefixes - Undocumented member.

viCustomerAddress - Undocumented member.

viVlan - Undocumented member.

viLocation - Undocumented member.

viAmazonAddress - Undocumented member.

viAddressFamily - Undocumented member.

viVirtualInterfaceState - Undocumented member.

viConnectionId - Undocumented member.

viVirtualInterfaceType - Undocumented member.

viAsn - Undocumented member.

viAuthKey - Undocumented member.

viCustomerRouterConfig - Information for generating the customer router configuration.

viOwnerAccount - The AWS account that will own the new virtual interface.

viVirtualInterfaceName - Undocumented member.

viVirtualInterfaceId - Undocumented member.

data VirtualInterface

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

See: virtualInterface smart constructor.

Response Lenses

viBgpPeers :: Lens' VirtualInterface [BGPPeer]

Undocumented member.

viVirtualGatewayId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viRouteFilterPrefixes :: Lens' VirtualInterface [RouteFilterPrefix]

Undocumented member.

viCustomerAddress :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVlan :: Lens' VirtualInterface (Maybe Int)

Undocumented member.

viLocation :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAmazonAddress :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAddressFamily :: Lens' VirtualInterface (Maybe AddressFamily)

Undocumented member.

viVirtualInterfaceState :: Lens' VirtualInterface (Maybe VirtualInterfaceState)

Undocumented member.

viConnectionId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVirtualInterfaceType :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAsn :: Lens' VirtualInterface (Maybe Int)

Undocumented member.

viAuthKey :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viCustomerRouterConfig :: Lens' VirtualInterface (Maybe Text)

Information for generating the customer router configuration.

viOwnerAccount :: Lens' VirtualInterface (Maybe Text)

The AWS account that will own the new virtual interface.

viVirtualInterfaceName :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVirtualInterfaceId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.