Network.AWS.DirectConnect.AllocatePrivateVirtualInterface

Description

Provisions a private virtual interface to be owned by a different customer.

The owner of a connection calls this function to provision a private virtual interface which will be owned by another AWS customer.

Virtual interfaces created using this function must be confirmed by the virtual interface owner by calling ConfirmPrivateVirtualInterface. Until this step has been completed, the virtual interface will be in Confirming state, and will not be available for handling traffic.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_AllocatePrivateVirtualInterface.html

Request

data AllocatePrivateVirtualInterface

Request constructor

allocatePrivateVirtualInterface

Arguments

:: Text
apviConnectionId

-> Text
apviOwnerAccount

-> NewPrivateVirtualInterfaceAllocation
apviNewPrivateVirtualInterfaceAllocation

-> AllocatePrivateVirtualInterface

AllocatePrivateVirtualInterface constructor.

The fields accessible through corresponding lenses are:

apviConnectionId :: Text

apviNewPrivateVirtualInterfaceAllocation :: NewPrivateVirtualInterfaceAllocation

apviOwnerAccount :: Text

Request lenses

apviConnectionId :: Lens' AllocatePrivateVirtualInterface Text

The connection ID on which the private virtual interface is provisioned.

Default: None

apviNewPrivateVirtualInterfaceAllocation :: Lens' AllocatePrivateVirtualInterface NewPrivateVirtualInterfaceAllocation

Detailed information for the private virtual interface to be provisioned.

Default: None

apviOwnerAccount :: Lens' AllocatePrivateVirtualInterface Text

The AWS account that will own the new private virtual interface.

Default: None

Response

data AllocatePrivateVirtualInterfaceResponse

Response constructor

allocatePrivateVirtualInterfaceResponse :: AllocatePrivateVirtualInterfaceResponse

AllocatePrivateVirtualInterfaceResponse constructor.

The fields accessible through corresponding lenses are:

apvirAmazonAddress :: Maybe Text

apvirAsn :: Maybe Int

apvirAuthKey :: Maybe Text

apvirConnectionId :: Maybe Text

apvirCustomerAddress :: Maybe Text

apvirCustomerRouterConfig :: Maybe Text

apvirLocation :: Maybe Text

apvirOwnerAccount :: Maybe Text

apvirRouteFilterPrefixes :: [RouteFilterPrefix]

apvirVirtualGatewayId :: Maybe Text

apvirVirtualInterfaceId :: Maybe Text

apvirVirtualInterfaceName :: Maybe Text

apvirVirtualInterfaceState :: Maybe VirtualInterfaceState

apvirVirtualInterfaceType :: Maybe Text

apvirVlan :: Maybe Int

Response lenses

apvirAmazonAddress :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirAsn :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Int)

apvirAuthKey :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirConnectionId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirCustomerAddress :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirCustomerRouterConfig :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

Information for generating the customer router configuration.

apvirLocation :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirOwnerAccount :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirRouteFilterPrefixes :: Lens' AllocatePrivateVirtualInterfaceResponse [RouteFilterPrefix]

apvirVirtualGatewayId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirVirtualInterfaceId :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirVirtualInterfaceName :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirVirtualInterfaceState :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

apvirVirtualInterfaceType :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Text)

apvirVlan :: Lens' AllocatePrivateVirtualInterfaceResponse (Maybe Int)