Network.AWS.DirectConnect.ConfirmConnection

Description

Confirm the creation of a hosted connection on an interconnect.

Upon creation, the hosted connection is initially in the Ordering state, and will remain in this state until the owner calls ConfirmConnection to confirm creation of the hosted connection.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_ConfirmConnection.html

Request

data ConfirmConnection

Request constructor

confirmConnection

Arguments

:: Text
ccConnectionId

-> ConfirmConnection

ConfirmConnection constructor.

The fields accessible through corresponding lenses are:

ccConnectionId :: Text

Request lenses

ccConnectionId :: Lens' ConfirmConnection Text

Response

data ConfirmConnectionResponse

Response constructor

confirmConnectionResponse :: ConfirmConnectionResponse

ConfirmConnectionResponse constructor.

The fields accessible through corresponding lenses are:

ccr1ConnectionState :: Maybe ConnectionState

Response lenses

ccr1ConnectionState :: Lens' ConfirmConnectionResponse (Maybe ConnectionState)