Network.AWS.DirectConnect.ConfirmPrivateVirtualInterface

Description

Accept ownership of a private virtual interface created by another customer.

After the virtual interface owner calls this function, the virtual interface will be created and attached to the given virtual private gateway, and will be available for handling traffic.

Creating a Request

confirmPrivateVirtualInterface

Arguments

:: Text
cpviVirtualInterfaceId

-> Text
cpviVirtualGatewayId

-> ConfirmPrivateVirtualInterface

Creates a value of ConfirmPrivateVirtualInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpviVirtualInterfaceId - Undocumented member.

cpviVirtualGatewayId - ID of the virtual private gateway that will be attached to the virtual interface. A virtual private gateway can be managed via the Amazon Virtual Private Cloud (VPC) console or the EC2 CreateVpnGateway action. Default: None

data ConfirmPrivateVirtualInterface

Container for the parameters to the ConfirmPrivateVirtualInterface operation.

See: confirmPrivateVirtualInterface smart constructor.

Request Lenses

cpviVirtualInterfaceId :: Lens' ConfirmPrivateVirtualInterface Text

Undocumented member.

cpviVirtualGatewayId :: Lens' ConfirmPrivateVirtualInterface Text

ID of the virtual private gateway that will be attached to the virtual interface. A virtual private gateway can be managed via the Amazon Virtual Private Cloud (VPC) console or the EC2 CreateVpnGateway action. Default: None

Destructuring the Response

confirmPrivateVirtualInterfaceResponse

Arguments

:: Int
cpvirsResponseStatus

-> ConfirmPrivateVirtualInterfaceResponse

Creates a value of ConfirmPrivateVirtualInterfaceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpvirsVirtualInterfaceState - Undocumented member.

cpvirsResponseStatus - -- | The response status code.

data ConfirmPrivateVirtualInterfaceResponse

The response received when ConfirmPrivateVirtualInterface is called.

See: confirmPrivateVirtualInterfaceResponse smart constructor.

Response Lenses

cpvirsVirtualInterfaceState :: Lens' ConfirmPrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

Undocumented member.

cpvirsResponseStatus :: Lens' ConfirmPrivateVirtualInterfaceResponse Int

- | The response status code.