Network.AWS.DirectConnect.CreateConnection

Description

Creates a new connection between the customer network and a specific AWS Direct Connect location.

A connection links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. An AWS Direct Connect location provides access to Amazon Web Services in the region it is associated with. You can establish connections with AWS Direct Connect locations in multiple regions, but a connection in one region does not provide connectivity to other regions.

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_CreateConnection.html

Request

data CreateConnection

Request constructor

createConnection

Arguments

:: Text
ccLocation

-> Text
ccBandwidth

-> Text
ccConnectionName

-> CreateConnection

CreateConnection constructor.

The fields accessible through corresponding lenses are:

ccBandwidth :: Text

ccConnectionName :: Text

ccLocation :: Text

Request lenses

ccBandwidth :: Lens' CreateConnection Text

ccConnectionName :: Lens' CreateConnection Text

ccLocation :: Lens' CreateConnection Text

Response

data CreateConnectionResponse

Response constructor

createConnectionResponse :: CreateConnectionResponse

CreateConnectionResponse constructor.

The fields accessible through corresponding lenses are:

ccrBandwidth :: Maybe Text

ccrConnectionId :: Maybe Text

ccrConnectionName :: Maybe Text

ccrConnectionState :: Maybe ConnectionState

ccrLocation :: Maybe Text

ccrOwnerAccount :: Maybe Text

ccrPartnerName :: Maybe Text

ccrRegion :: Maybe Text

ccrVlan :: Maybe Int

Response lenses

ccrBandwidth :: Lens' CreateConnectionResponse (Maybe Text)

Bandwidth of the connection.

Example: 1Gbps (for regular connections), or 500Mbps (for hosted connections)

Default: None

ccrConnectionId :: Lens' CreateConnectionResponse (Maybe Text)

ccrConnectionName :: Lens' CreateConnectionResponse (Maybe Text)

ccrConnectionState :: Lens' CreateConnectionResponse (Maybe ConnectionState)

ccrLocation :: Lens' CreateConnectionResponse (Maybe Text)

ccrOwnerAccount :: Lens' CreateConnectionResponse (Maybe Text)

ccrPartnerName :: Lens' CreateConnectionResponse (Maybe Text)

ccrRegion :: Lens' CreateConnectionResponse (Maybe Text)

ccrVlan :: Lens' CreateConnectionResponse (Maybe Int)