Network.AWS.DirectConnect.CreatePrivateVirtualInterface

Description

Creates a new private virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A private virtual interface supports sending traffic to a single virtual private cloud (VPC).

Creating a Request

createPrivateVirtualInterface

Arguments

:: Text
creConnectionId

-> NewPrivateVirtualInterface
creNewPrivateVirtualInterface

-> CreatePrivateVirtualInterface

Creates a value of CreatePrivateVirtualInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

creConnectionId - Undocumented member.

creNewPrivateVirtualInterface - Detailed information for the private virtual interface to be created. Default: None

data CreatePrivateVirtualInterface

Container for the parameters to the CreatePrivateVirtualInterface operation.

See: createPrivateVirtualInterface smart constructor.

Request Lenses

creConnectionId :: Lens' CreatePrivateVirtualInterface Text

Undocumented member.

creNewPrivateVirtualInterface :: Lens' CreatePrivateVirtualInterface NewPrivateVirtualInterface

Detailed information for the private virtual interface to be created. Default: None

Destructuring the Response

virtualInterface :: VirtualInterface

Creates a value of VirtualInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

viBgpPeers - Undocumented member.

viVirtualGatewayId - Undocumented member.

viRouteFilterPrefixes - Undocumented member.

viCustomerAddress - Undocumented member.

viVlan - Undocumented member.

viLocation - Undocumented member.

viAmazonAddress - Undocumented member.

viAddressFamily - Undocumented member.

viVirtualInterfaceState - Undocumented member.

viConnectionId - Undocumented member.

viVirtualInterfaceType - Undocumented member.

viAsn - Undocumented member.

viAuthKey - Undocumented member.

viCustomerRouterConfig - Information for generating the customer router configuration.

viOwnerAccount - The AWS account that will own the new virtual interface.

viVirtualInterfaceName - Undocumented member.

viVirtualInterfaceId - Undocumented member.

data VirtualInterface

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

See: virtualInterface smart constructor.

Response Lenses

viBgpPeers :: Lens' VirtualInterface [BGPPeer]

Undocumented member.

viVirtualGatewayId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viRouteFilterPrefixes :: Lens' VirtualInterface [RouteFilterPrefix]

Undocumented member.

viCustomerAddress :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVlan :: Lens' VirtualInterface (Maybe Int)

Undocumented member.

viLocation :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAmazonAddress :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAddressFamily :: Lens' VirtualInterface (Maybe AddressFamily)

Undocumented member.

viVirtualInterfaceState :: Lens' VirtualInterface (Maybe VirtualInterfaceState)

Undocumented member.

viConnectionId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVirtualInterfaceType :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viAsn :: Lens' VirtualInterface (Maybe Int)

Undocumented member.

viAuthKey :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viCustomerRouterConfig :: Lens' VirtualInterface (Maybe Text)

Information for generating the customer router configuration.

viOwnerAccount :: Lens' VirtualInterface (Maybe Text)

The AWS account that will own the new virtual interface.

viVirtualInterfaceName :: Lens' VirtualInterface (Maybe Text)

Undocumented member.

viVirtualInterfaceId :: Lens' VirtualInterface (Maybe Text)

Undocumented member.