Network.AWS.DirectConnect.CreatePrivateVirtualInterface

Description

Creates a new private virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A private virtual interface supports sending traffic to a single virtual private cloud (VPC).

http://docs.aws.amazon.com/directconnect/latest/APIReference/API_CreatePrivateVirtualInterface.html

Request

data CreatePrivateVirtualInterface

Request constructor

createPrivateVirtualInterface

Arguments

:: Text
cpvi1ConnectionId

-> NewPrivateVirtualInterface
cpvi1NewPrivateVirtualInterface

-> CreatePrivateVirtualInterface

CreatePrivateVirtualInterface constructor.

The fields accessible through corresponding lenses are:

cpvi1ConnectionId :: Text

cpvi1NewPrivateVirtualInterface :: NewPrivateVirtualInterface

Request lenses

cpvi1ConnectionId :: Lens' CreatePrivateVirtualInterface Text

cpvi1NewPrivateVirtualInterface :: Lens' CreatePrivateVirtualInterface NewPrivateVirtualInterface

Detailed information for the private virtual interface to be created.

Default: None

Response

data CreatePrivateVirtualInterfaceResponse

Response constructor

createPrivateVirtualInterfaceResponse :: CreatePrivateVirtualInterfaceResponse

CreatePrivateVirtualInterfaceResponse constructor.

The fields accessible through corresponding lenses are:

cpvir2AmazonAddress :: Maybe Text

cpvir2Asn :: Maybe Int

cpvir2AuthKey :: Maybe Text

cpvir2ConnectionId :: Maybe Text

cpvir2CustomerAddress :: Maybe Text

cpvir2CustomerRouterConfig :: Maybe Text

cpvir2Location :: Maybe Text

cpvir2OwnerAccount :: Maybe Text

cpvir2RouteFilterPrefixes :: [RouteFilterPrefix]

cpvir2VirtualGatewayId :: Maybe Text

cpvir2VirtualInterfaceId :: Maybe Text

cpvir2VirtualInterfaceName :: Maybe Text

cpvir2VirtualInterfaceState :: Maybe VirtualInterfaceState

cpvir2VirtualInterfaceType :: Maybe Text

cpvir2Vlan :: Maybe Int

Response lenses

cpvir2AmazonAddress :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir2Asn :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Int)

cpvir2AuthKey :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir2ConnectionId :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir2CustomerAddress :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir2CustomerRouterConfig :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

Information for generating the customer router configuration.

cpvir2Location :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir2OwnerAccount :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir2RouteFilterPrefixes :: Lens' CreatePrivateVirtualInterfaceResponse [RouteFilterPrefix]

cpvir2VirtualGatewayId :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir2VirtualInterfaceId :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir2VirtualInterfaceName :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir2VirtualInterfaceState :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

cpvir2VirtualInterfaceType :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir2Vlan :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Int)