Network.AWS.DirectConnect.DescribeVirtualInterfaces

Description

Displays all virtual interfaces for an AWS account. Virtual interfaces deleted fewer than 15 minutes before you make the request are also returned. If you specify a connection ID, only the virtual interfaces associated with the connection are returned. If you specify a virtual interface ID, then only a single virtual interface is returned.

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

Creating a Request

describeVirtualInterfaces :: DescribeVirtualInterfaces

Creates a value of DescribeVirtualInterfaces with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dviConnectionId - Undocumented member.

dviVirtualInterfaceId - Undocumented member.

data DescribeVirtualInterfaces

Container for the parameters to the DescribeVirtualInterfaces operation.

See: describeVirtualInterfaces smart constructor.

Request Lenses

dviConnectionId :: Lens' DescribeVirtualInterfaces (Maybe Text)

Undocumented member.

dviVirtualInterfaceId :: Lens' DescribeVirtualInterfaces (Maybe Text)

Undocumented member.

Destructuring the Response

describeVirtualInterfacesResponse

Arguments

:: Int
dvisrsResponseStatus

-> DescribeVirtualInterfacesResponse

Creates a value of DescribeVirtualInterfacesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvisrsVirtualInterfaces - A list of virtual interfaces.

dvisrsResponseStatus - -- | The response status code.

data DescribeVirtualInterfacesResponse

A structure containing a list of virtual interfaces.

See: describeVirtualInterfacesResponse smart constructor.

Response Lenses

dvisrsVirtualInterfaces :: Lens' DescribeVirtualInterfacesResponse [VirtualInterface]

A list of virtual interfaces.

dvisrsResponseStatus :: Lens' DescribeVirtualInterfacesResponse Int

- | The response status code.