Network.AWS.DirectConnect.V2012_10_25.ConfirmPrivateVirtualInterface

Description

Accept ownership of a private virtual interface created by another customer. After the virtual interface owner calls this function, the virtual interface will be created and attached to the given virtual private gateway, and will be available for handling traffic.

Request

data ConfirmPrivateVirtualInterface

Container for the parameters to the ConfirmPrivateVirtualInterface operation.

Request constructor

mkConfirmPrivateVirtualInterface

Arguments

:: Text
cpviVirtualInterfaceId

-> Text
cpviVirtualGatewayId

-> ConfirmPrivateVirtualInterface

Smart constructor for the minimum required parameters to construct a valid ConfirmPrivateVirtualInterface request.

Request lenses

cpviVirtualInterfaceId :: Lens' ConfirmPrivateVirtualInterface Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

cpviVirtualGatewayId :: Lens' ConfirmPrivateVirtualInterface Text

ID of the virtual private gateway that will be attached to the virtual interface. A virtual private gateway can be managed via the Amazon Virtual Private Cloud (VPC) console or the EC2 CreateVpnGateway action. Default: None.

Response

data ConfirmPrivateVirtualInterfaceResponse

The response received when ConfirmPrivateVirtualInterface is called.

Response lenses

cpvirVirtualInterfaceState :: Lens' ConfirmPrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.