Network.AWS.DirectConnect.V2012_10_25.CreatePrivateVirtualInterface

Description

Creates a new private virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A private virtual interface supports sending traffic to a single virtual private cloud (VPC).

Request

data CreatePrivateVirtualInterface

Container for the parameters to the CreatePrivateVirtualInterface operation.

Request constructor

mkCreatePrivateVirtualInterface

Arguments

:: Text
cpvi2ConnectionId

-> NewPrivateVirtualInterface
cpvi2NewPrivateVirtualInterface

-> CreatePrivateVirtualInterface

Smart constructor for the minimum required parameters to construct a valid CreatePrivateVirtualInterface request.

Request lenses

cpvi2ConnectionId :: Lens' CreatePrivateVirtualInterface Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

cpvi2NewPrivateVirtualInterface :: Lens' CreatePrivateVirtualInterface NewPrivateVirtualInterface

Detailed information for the private virtual interface to be created. Default: None.

Response

data CreatePrivateVirtualInterfaceResponse

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

Response lenses

cpvir1OwnerAccount :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

cpvir1VirtualInterfaceId :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

cpvir1Location :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

Where the connection is located. Example: EqSV5 Default: None.

cpvir1ConnectionId :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

ID of the connection. Example: dxcon-fg5678gh Default: None.

cpvir1VirtualInterfaceType :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

cpvir1VirtualInterfaceName :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

The name of the virtual interface assigned by the customer. Example: "My VPC".

cpvir1Vlan :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Integer)

The VLAN ID. Example: 101.

cpvir1Asn :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Integer)

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

cpvir1AuthKey :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

Authentication key for BGP configuration. Example: asdf34example.

cpvir1AmazonAddress :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

cpvir1CustomerAddress :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

IP address assigned to the customer interface. Example: 192.168.1.2/30.

cpvir1VirtualInterfaceState :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe VirtualInterfaceState)

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

cpvir1CustomerRouterConfig :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

Information for generating the customer router configuration.

cpvir1VirtualGatewayId :: Lens' CreatePrivateVirtualInterfaceResponse (Maybe Text)

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

cpvir1RouteFilterPrefixes :: Lens' CreatePrivateVirtualInterfaceResponse [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).