Network.AWS.DirectConnect.V2012_10_25.DescribeInterconnects

Description

Returns a list of interconnects owned by the AWS account. If an interconnect ID is provided, it will only return this particular interconnect.

Request

data DescribeInterconnects

Container for the parameters to the DescribeInterconnects operation.

Request constructor

mkDescribeInterconnects :: DescribeInterconnects

Smart constructor for the minimum required parameters to construct a valid DescribeInterconnects request.

Request lenses

di1InterconnectId :: Lens' DescribeInterconnects (Maybe Text)

The ID of the interconnect. Example: dxcon-abc123.

Response

data DescribeInterconnectsResponse

A structure containing a list of interconnects.

Response constructor

mkDescribeInterconnectsResponse :: DescribeInterconnectsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeInterconnectsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

dirrInterconnects :: Lens' DescribeInterconnectsResponse [Interconnect]

A list of interconnects.