Network.AWS.DirectConnect.V2012_10_25.Types

Description

AWS Direct Connect links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. With this connection in place, you can create virtual interfaces directly to the AWS cloud and Amazon Virtual Private Cloud, bypassing Internet service providers in your network path.

Documentation

data DirectConnect

Supported version (2012-10-25) of the AWS Direct Connect service.

data ConnectionState

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

Constructors

ConnectionStateAvailable
available

ConnectionStateDeleted
deleted

ConnectionStateDeleting
deleting

ConnectionStateDown
down

ConnectionStateOrdering
ordering

ConnectionStatePending
pending

ConnectionStateRejected
rejected

ConnectionStateRequested
requested

data InterconnectState

State of the interconnect. Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The interconnect has been approved, and is being initialized. Available: The network link is up, and the interconnect is ready for use. Down: The network link is down. Deleted: The interconnect has been deleted.

Constructors

InterconnectStateAvailable
available

InterconnectStateDeleted
deleted

InterconnectStateDeleting
deleting

InterconnectStateDown
down

InterconnectStatePending
pending

InterconnectStateRequested
requested

data VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

Constructors

VirtualInterfaceStateAvailable
available

VirtualInterfaceStateConfirming
confirming

VirtualInterfaceStateDeleted
deleted

VirtualInterfaceStateDeleting
deleting

VirtualInterfaceStatePending
pending

VirtualInterfaceStateRejected
rejected

VirtualInterfaceStateVerifying
verifying

newtype RouteFilterPrefix

A route filter prefix that the customer can advertise through Border Gateway Protocol (BGP) over a public virtual interface.

Constructors

RouteFilterPrefix

Fields

_rfpCidr :: Maybe Text

CIDR notation for the advertised route. Multiple routes are separated by commas. Example: 10.10.10.024,10.10.11.024.

data Connection

A connection represents the physical network connection between the AWS Direct Connect location and the customer.

Constructors

Connection

Fields

_kVlan :: Maybe Integer

The VLAN ID. Example: 101.

_kLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_kConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_kPartnerName :: Maybe Text

_kConnectionName :: Maybe Text

The name of the connection. Example: "1G Connection to AWS" Default: None.

_kBandwidth :: Maybe Text

Bandwidth of the connection. Example: 1Gbps Default: None.

_kOwnerAccount :: Maybe Text

_kRegion :: Maybe Text

The AWS region where the connection is located. Example: us-east-1 Default: None.

_kConnectionState :: Maybe ConnectionState

State of the connection. Ordering: The initial state of a hosted connection provisioned on an interconnect. The connection stays in the ordering state until the owner of the hosted connection confirms or declines the connection order. Requested: The initial state of a standard connection. The connection stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The connection has been approved, and is being initialized. Available: The network link is up, and the connection is ready for use. Down: The network link is down. Deleted: The connection has been deleted. Rejected: A hosted connection in the Ordering state will enter the Rejected state if it is deleted by the end customer.

data Interconnect

An interconnect is a connection that can host other connections. Like a standard AWS Direct Connect connection, an interconnect represents the physical connection between an AWS Direct Connect partner's network and a specific Direct Connect location. An AWS Direct Connect partner who owns an interconnect can provision hosted connections on the interconnect for their end customers, thereby providing the end customers with connectivity to AWS services. The resources of the interconnect, including bandwidth and VLAN numbers, are shared by all of the hosted connections on the interconnect, and the owner of the interconnect determines how these resources are assigned.

Constructors

Interconnect

Fields

_jInterconnectId :: Maybe Text

The ID of the interconnect. Example: dxcon-abc123.

_jLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_jInterconnectName :: Maybe Text

The name of the interconnect. Example: "1G Interconnect to AWS".

_jBandwidth :: Maybe Text

Bandwidth of the connection. Example: 1Gbps Default: None.

_jInterconnectState :: Maybe InterconnectState

State of the interconnect. Requested: The initial state of an interconnect. The interconnect stays in the requested state until the Letter of Authorization (LOA) is sent to the customer. Pending: The interconnect has been approved, and is being initialized. Available: The network link is up, and the interconnect is ready for use. Down: The network link is down. Deleted: The interconnect has been deleted.

_jRegion :: Maybe Text

The AWS region where the connection is located. Example: us-east-1 Default: None.

data Location

An AWS Direct Connect location where connections and interconnects can be requested.

Constructors

Location

Fields

_qLocationName :: Maybe Text

The name of the AWS Direct Connect location. The name includes the colocation partner name and the physical site of the lit building.

_qLocationCode :: Maybe Text

The code used to indicate the AWS Direct Connect location.

data NewPrivateVirtualInterface

Detailed information for the private virtual interface to be created. Default: None.

Constructors

NewPrivateVirtualInterface

Fields

_npvjVirtualGatewayId :: Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_npvjCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_npvjVlan :: Integer

The VLAN ID. Example: 101.

_npvjAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_npvjAsn :: Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_npvjAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_npvjVirtualInterfaceName :: Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

data NewPrivateVirtualInterfaceAllocation

Detailed information for the private virtual interface to be provisioned. Default: None.

Constructors

NewPrivateVirtualInterfaceAllocation

Fields

_npviaCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_npviaVlan :: Integer

The VLAN ID. Example: 101.

_npviaAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_npviaAsn :: Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_npviaAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_npviaVirtualInterfaceName :: Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

data NewPublicVirtualInterface

Detailed information for the public virtual interface to be created. Default: None.

Constructors

NewPublicVirtualInterface

Fields

_npviRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_npviCustomerAddress :: Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_npviVlan :: Integer

The VLAN ID. Example: 101.

_npviAmazonAddress :: Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_npviAsn :: Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_npviAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_npviVirtualInterfaceName :: Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

data NewPublicVirtualInterfaceAllocation

Detailed information for the public virtual interface to be provisioned. Default: None.

Constructors

NewPublicVirtualInterfaceAllocation

Fields

_npvibRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_npvibCustomerAddress :: Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_npvibVlan :: Integer

The VLAN ID. Example: 101.

_npvibAmazonAddress :: Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_npvibAsn :: Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_npvibAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_npvibVirtualInterfaceName :: Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

data VirtualGateway

You can create one or more AWS Direct Connect private virtual interfaces linking to your virtual private gateway. Virtual private gateways can be managed using the Amazon Virtual Private Cloud (Amazon VPC) console or the Amazon EC2 CreateVpnGateway action.

Constructors

VirtualGateway

Fields

_vhVirtualGatewayId :: Maybe Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_vhVirtualGatewayState :: Maybe Text

State of the virtual private gateway. Pending: This is the initial state after calling CreateVpnGateway. Available: Ready for use by a private virtual interface. Deleting: This is the initial state after calling DeleteVpnGateway. Deleted: In this state, a private virtual interface is unable to send traffic over this gateway.

data VirtualInterface

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

Constructors

VirtualInterface

Fields

_vlVirtualGatewayId :: Maybe Text

The ID of the virtual private gateway to a VPC. This only applies to private virtual interfaces. Example: vgw-123er56.

_vlRouteFilterPrefixes :: [RouteFilterPrefix]

A list of routes to be advertised to the AWS network in this region (public virtual interface) or your VPC (private virtual interface).

_vlCustomerAddress :: Maybe Text

IP address assigned to the customer interface. Example: 192.168.1.2/30.

_vlVlan :: Maybe Integer

The VLAN ID. Example: 101.

_vlLocation :: Maybe Text

Where the connection is located. Example: EqSV5 Default: None.

_vlAmazonAddress :: Maybe Text

IP address assigned to the Amazon interface. Example: 192.168.1.1/30.

_vlVirtualInterfaceState :: Maybe VirtualInterfaceState

State of the virtual interface. Confirming: The creation of the virtual interface is pending confirmation from the virtual interface owner. If the owner of the virtual interface is different from the owner of the connection on which it is provisioned, then the virtual interface will remain in this state until it is confirmed by the virtual interface owner. Verifying: This state only applies to public virtual interfaces. Each public virtual interface needs validation before the virtual interface can be created. Pending: A virtual interface is in this state from the time that it is created until the virtual interface is ready to forward traffic. Available: A virtual interface that is able to forward traffic. Deleting: A virtual interface is in this state immediately after calling DeleteVirtualInterface until it can no longer forward traffic. Deleted: A virtual interface that cannot forward traffic. Rejected: The virtual interface owner has declined creation of the virtual interface. If a virtual interface in the Confirming state is deleted by the virtual interface owner, the virtual interface will enter the Rejected state.

_vlConnectionId :: Maybe Text

ID of the connection. Example: dxcon-fg5678gh Default: None.

_vlVirtualInterfaceType :: Maybe Text

The type of virtual interface. Example: private (Amazon VPC) or public (Amazon S3, Amazon DynamoDB, and so on.).

_vlAsn :: Maybe Integer

Autonomous system (AS) number for Border Gateway Protocol (BGP) configuration. Example: 65000.

_vlAuthKey :: Maybe Text

Authentication key for BGP configuration. Example: asdf34example.

_vlCustomerRouterConfig :: Maybe Text

Information for generating the customer router configuration.

_vlOwnerAccount :: Maybe Text

_vlVirtualInterfaceName :: Maybe Text

The name of the virtual interface assigned by the customer. Example: "My VPC".

_vlVirtualInterfaceId :: Maybe Text

ID of the virtual interface. Example: dxvif-123dfg56 Default: None.

rfpCidr :: Iso' RouteFilterPrefix (Maybe Text)

kVlan :: Lens' Connection (Maybe Integer)

kRegion :: Lens' Connection (Maybe Text)

kPartnerName :: Lens' Connection (Maybe Text)

kOwnerAccount :: Lens' Connection (Maybe Text)

kLocation :: Lens' Connection (Maybe Text)

kConnectionState :: Lens' Connection (Maybe ConnectionState)

kConnectionName :: Lens' Connection (Maybe Text)

kConnectionId :: Lens' Connection (Maybe Text)

kBandwidth :: Lens' Connection (Maybe Text)

jRegion :: Lens' Interconnect (Maybe Text)

jLocation :: Lens' Interconnect (Maybe Text)

jInterconnectState :: Lens' Interconnect (Maybe InterconnectState)

jInterconnectName :: Lens' Interconnect (Maybe Text)

jInterconnectId :: Lens' Interconnect (Maybe Text)

jBandwidth :: Lens' Interconnect (Maybe Text)

qLocationName :: Lens' Location (Maybe Text)

qLocationCode :: Lens' Location (Maybe Text)

npvjVlan :: Lens' NewPrivateVirtualInterface Integer

npvjVirtualInterfaceName :: Lens' NewPrivateVirtualInterface Text

npvjVirtualGatewayId :: Lens' NewPrivateVirtualInterface Text

npvjCustomerAddress :: Lens' NewPrivateVirtualInterface (Maybe Text)

npvjAuthKey :: Lens' NewPrivateVirtualInterface (Maybe Text)

npvjAsn :: Lens' NewPrivateVirtualInterface Integer

npvjAmazonAddress :: Lens' NewPrivateVirtualInterface (Maybe Text)

npviaVlan :: Lens' NewPrivateVirtualInterfaceAllocation Integer

npviaVirtualInterfaceName :: Lens' NewPrivateVirtualInterfaceAllocation Text

npviaCustomerAddress :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

npviaAuthKey :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

npviaAsn :: Lens' NewPrivateVirtualInterfaceAllocation Integer

npviaAmazonAddress :: Lens' NewPrivateVirtualInterfaceAllocation (Maybe Text)

npviVlan :: Lens' NewPublicVirtualInterface Integer

npviVirtualInterfaceName :: Lens' NewPublicVirtualInterface Text

npviRouteFilterPrefixes :: Lens' NewPublicVirtualInterface [RouteFilterPrefix]

npviCustomerAddress :: Lens' NewPublicVirtualInterface Text

npviAuthKey :: Lens' NewPublicVirtualInterface (Maybe Text)

npviAsn :: Lens' NewPublicVirtualInterface Integer

npviAmazonAddress :: Lens' NewPublicVirtualInterface Text

npvibVlan :: Lens' NewPublicVirtualInterfaceAllocation Integer

npvibVirtualInterfaceName :: Lens' NewPublicVirtualInterfaceAllocation Text

npvibRouteFilterPrefixes :: Lens' NewPublicVirtualInterfaceAllocation [RouteFilterPrefix]

npvibCustomerAddress :: Lens' NewPublicVirtualInterfaceAllocation Text

npvibAuthKey :: Lens' NewPublicVirtualInterfaceAllocation (Maybe Text)

npvibAsn :: Lens' NewPublicVirtualInterfaceAllocation Integer

npvibAmazonAddress :: Lens' NewPublicVirtualInterfaceAllocation Text

vhVirtualGatewayState :: Lens' VirtualGateway (Maybe Text)

vhVirtualGatewayId :: Lens' VirtualGateway (Maybe Text)

vlVlan :: Lens' VirtualInterface (Maybe Integer)

vlVirtualInterfaceType :: Lens' VirtualInterface (Maybe Text)

vlVirtualInterfaceState :: Lens' VirtualInterface (Maybe VirtualInterfaceState)

vlVirtualInterfaceName :: Lens' VirtualInterface (Maybe Text)

vlVirtualInterfaceId :: Lens' VirtualInterface (Maybe Text)

vlVirtualGatewayId :: Lens' VirtualInterface (Maybe Text)

vlRouteFilterPrefixes :: Lens' VirtualInterface [RouteFilterPrefix]

vlOwnerAccount :: Lens' VirtualInterface (Maybe Text)

vlLocation :: Lens' VirtualInterface (Maybe Text)

vlCustomerRouterConfig :: Lens' VirtualInterface (Maybe Text)

vlCustomerAddress :: Lens' VirtualInterface (Maybe Text)

vlConnectionId :: Lens' VirtualInterface (Maybe Text)

vlAuthKey :: Lens' VirtualInterface (Maybe Text)

vlAsn :: Lens' VirtualInterface (Maybe Integer)

vlAmazonAddress :: Lens' VirtualInterface (Maybe Text)